% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_around_decimal_point.r
\name{aadp}
\alias{aadp}
\title{Pad Numeric Vector Around Decimal Point}
\usage{
aadp(x, sep = "\\\\.", len)
}
\arguments{
\item{x}{A vector of numbers.}

\item{sep}{The decimal separator.}

\item{len}{The total character width desired in the formatted field.}
}
\value{
A character vector with members of length \code{len}, with \code{sep} located in the same position
of each member.
}
\description{
Pads a vector of numeric values so that the decimal is at the same ordinal location.
}
\examples{
aadp(c(1,100, 0.01))
}
\seealso{
Other padding: 
\code{\link{align_around_decimal_point}()}
}
\concept{padding}
