% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cov.r
\name{as.list.mat}
\alias{as.list.mat}
\title{Extract a Matrix of a Converged NONMEM Run}
\usage{
\method{as.list}{mat}(x, ...)
}
\arguments{
\item{x}{'mat' object, e.g. output of \code{\link{read_cov.character}}}

\item{...}{ignored}
}
\value{
list of covariance matrices as data frames
}
\description{
Parses a 'mat' object (see \code{\link{read_cov}}) as a list 
of one data.frame for each estimation method
}
\seealso{
Other mat: 
\code{\link{as.matrix.mat}()}
}
\concept{mat}
