% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ext.r
\name{as.matrix.ext}
\alias{as.matrix.ext}
\title{Convert EXT to Matrix}
\usage{
\method{as.matrix}{ext}(x, ...)
}
\arguments{
\item{x}{ext}

\item{...}{ignored}
}
\value{
extList
}
\description{
Converts class 'ext' to matrix.
Alias for \code{\link{as.list.ext}}
}
\seealso{
Other ext: 
\code{\link{as.list.ext}()},
\code{\link{find_diag}()},
\code{\link{format.nmpartab}()},
\code{\link{format.nmpartab_}()},
\code{\link{get_ext}()},
\code{\link{get_ext.character}()},
\code{\link{get_ext.ext}()},
\code{\link{get_ext.extList}()},
\code{\link{get_ofv.ext}()},
\code{\link{get_ofv.nonmem}()},
\code{\link{lol}()},
\code{\link{nm_condition_number}()},
\code{\link{nm_condition_number.default}()},
\code{\link{nm_condition_number.ext}()},
\code{\link{nm_params_table}()},
\code{\link{nm_params_table.character}()},
\code{\link{nm_params_table_short}()},
\code{\link{nm_params_table_short.character}()},
\code{\link{nm_params_table_short.ext}()},
\code{\link{nm_params_table_short.nmpartab}()},
\code{\link{read_ext}()},
\code{\link{read_ext.character}()},
\code{\link{read_ext.ext}()},
\code{\link{summary.extList}()}
}
\concept{ext}
