% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cov.r
\name{as.matrix.mat}
\alias{as.matrix.mat}
\title{Extract a Matrix of a Converged NONMEM Run}
\usage{
\method{as.matrix}{mat}(x, ...)
}
\arguments{
\item{x}{'mat' object, e.g. output of \code{\link{read_cov.character}}}

\item{...}{ignored}
}
\value{
list of matrices as data frames
}
\description{
Parses a 'mat' object (see e.g. \code{\link{read_cov}}) as a list 
of one data.frame for each estimation method
}
\examples{

cov <- read_cov('example1', directory = getOption("qpExampleDir"))
covmat =  as.matrix(cov)
names(covmat)
names(covmat[[1]])
lapply(covmat, dim)

}
\seealso{
Other mat: 
\code{\link{as.list.mat}()}
}
\concept{mat}
