% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_categorical.R
\name{as_categorical}
\alias{as_categorical}
\title{Coerce to Categorical
Coerce to categorical. 
Generic, with method \code{\link{as_categorical.decorated}}.}
\usage{
as_categorical(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed}
}
\description{
Coerce to Categorical
Coerce to categorical. 
Generic, with method \code{\link{as_categorical.decorated}}.
}
\seealso{
Other decorated: 
\code{\link{as_categorical.decorated}()}
}
\concept{decorated}
\keyword{internal}
