% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tidyvpc_qp.r
\name{as_qpvpc}
\alias{as_qpvpc}
\title{Convert to qP style VPC}
\usage{
as_qpvpc(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed}
}
\description{
Converts to qP-style VPC.
Generic, with method \code{\link{as_qpvpc.tidyvpcobj}}.
}
\seealso{
Other tidyvpc: 
\code{\link{as_qpvpc.tidyvpcobj}()},
\code{\link{plot.qpvpc}()}
}
\concept{tidyvpc}
\keyword{internal}
