% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tidyvpc_qp.r
\name{as_qpvpc.tidyvpcobj}
\alias{as_qpvpc.tidyvpcobj}
\title{Convert 'tidyvpcobj' to 'qpvpc'}
\usage{
\method{as_qpvpc}{tidyvpcobj}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed}
}
\description{
Converts 'tidyvpcobj' to 'qpvpc'.
Simply supplies a new class (used for invoking custom print method).
}
\seealso{
Other tidyvpc: 
\code{\link{as_qpvpc}()},
\code{\link{plot.qpvpc}()}
}
\concept{tidyvpc}
