% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_params_table.r
\name{as_spork.nms_nonmem}
\alias{as_spork.nms_nonmem}
\title{Convert nms_nonmem to spork}
\usage{
\method{as_spork}{nms_nonmem}(x, ...)
}
\arguments{
\item{x}{nms_canonical; see \code{\link[nonmemica]{as_nms_canonical}}}

\item{...}{passed arguments}
}
\value{
spork
}
\description{
Converts nms_nonmem to spork, i.e. an abstract specification of 
thetas, omegas, and sigmas using NONMEM conventions.
}
