% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_classes.r
\name{compare_classes}
\alias{compare_classes}
\title{Compare Two Data Frames}
\usage{
compare_classes(df1, df2, ...)
}
\arguments{
\item{df1}{data.frame}

\item{df2}{data.frame}

\item{...}{ignored}
}
\value{
used for side effects
}
\description{
Compares the classes of two data.frames.
}
\examples{
library(magrittr)
library(dplyr)
library(tibble)
df1 = example_pkpd_data() \%>\% as_tibble
df2 = df1 \%>\% mutate(sex=ifelse(sex=="F",1,0), keo=as.character(keo))
compare_classes(df1,df2)
compare_classes(df1,df2) \%>\% filter(mark==1)
}
\seealso{
Other basic: 
\code{\link{asNumeric}()},
\code{\link{doubleStack}()},
\code{\link{doubleStack.decorated}()},
\code{\link{doubleStack.default}()},
\code{\link{extract_character}()},
\code{\link{extract_number}()},
\code{\link{fxs}()},
\code{\link{geocv}()},
\code{\link{geomean}()},
\code{\link{isMissing}()},
\code{\link{isNumeric}()},
\code{\link{logit}()},
\code{\link{logit_inv}()},
\code{\link{lunique}()},
\code{\link{make_symmetric}()},
\code{\link{odwg}()},
\code{\link{odwg_classification.data.frame}()},
\code{\link{rgenpois}()},
\code{\link{rnorm_by_id}()},
\code{\link{runc}()},
\code{\link{sample_by_id}()},
\code{\link{sunique}()}
}
\concept{basic}
