% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm_qp.r
\name{definitions.nonmem}
\alias{definitions.nonmem}
\title{Harvest Model Item Definitions for NONMEM}
\usage{
\method{definitions}{nonmem}(x, ...)
}
\arguments{
\item{x}{nonmem, i.e. output of \code{\link{read_nm_qp}}}

\item{...}{passed to \code{\link[nonmemica]{definitions}}}
}
\value{
definitions data.frame
}
\description{
Harvests model item definitions for NONMEM object.
}
\seealso{
read_nm_qp

Other nm: 
\code{\link{as_pmx_canonical}()},
\code{\link{extract_varcov}()},
\code{\link{get_est_table.default}()},
\code{\link{get_est_table.nonmem}()},
\code{\link{get_est_table}()},
\code{\link{nm_cov_table}()},
\code{\link{nm_covmat_extract}()},
\code{\link{nm_unzip}()},
\code{\link{read_nm_qp}()},
\code{\link{sample_uncert_qp}()}
}
\concept{nm}
