% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleStack.r
\name{doubleStack.decorated}
\alias{doubleStack.decorated}
\title{Stack a Value Twice by Two Covariates}
\usage{
\method{doubleStack}{decorated}(x, vars1, vars2, ...)
}
\arguments{
\item{x}{dataset containing all variables specified in vars1 and vars2}

\item{vars1}{The first variable to stack by. For example \code{c(WT,AGE,BMI,SEX)}}

\item{vars2}{The second variable to stack by after the first stack. For example \code{c(ETA1,ETA2,ETA3)}}

\item{...}{ignored}
}
\value{
A decorated data.frame
}
\description{
Dataset will be melted by specified columns after it will be melted by a second set of columns.
This comes in handy when plotting multiple individual parameter estimates versus multiple covariates.
This is the method for class "decorated".  See also \code{\link{doubleStack.default}}. The advantage here is that you can have variables that are not integer or numeric.  Also, you can have words rather than numeric values in the axes of resulting plots.  See examples.
}
\examples{
set.seed(0)
library(magrittr)
library(dplyr)
library(yamlet)
library(tidyr)
library(ggplot2)
x <- data.frame(id = 1:100)
x \%<>\% mutate(wt  = id \%>\% rnorm_by_id(76, 15) \%>\% signif)
x \%<>\% mutate(age = id \%>\% sample_by_id(18:99))
x \%<>\% mutate(sex = id \%>\% sample_by_id(0:1))
x \%<>\% mutate(coh = id \%>\% sample_by_id(1:3))
x \%<>\% mutate(CL  = id \%>\% rnorm_by_id(10, 1) \%>\% signif)
x \%<>\% mutate(V   = id \%>\% rnorm_by_id(3, 0.25) \%>\% signif)

# At the moment, we just have a data.frame, that responds normally 
y <- doubleStack(x, vars1 = c("CL","V"), vars2 = c("wt","age","sex"))

# Now we decorate it.
head(x)
x \%<>\% decorate('
id: Subject ID
wt:  [ Body Weight, kg]
age: [ Age, year]
sex: [ Sex, [ Female: 0, Male: 1 ]]
coh: [ Cohort, [ Cohort 1: 1, Cohort 2: 2, Cohort 3: 3]]
CL:  [ CL/F, L/h]
V:   [ V/F, L ]
')

# Select some variables
vars1 = c('CL', 'V')
vars2 = c('coh', 'sex')

# A conventional plot gives numbers in x axis labels
x \%>\% 
   doubleStack(vars1, vars2) \%>\%
   mutate(value2 = factor(value2)) \%>\%
   ggplot(aes(x = value2, y = value1)) +
   facet_grid(variable1 ~ variable2, scales = 'free') +
   geom_boxplot()

# A plot of "resolved" data is more informative.
x \%>\% 
   resolve \%>\%
   modify(coh, sex, name = label) \%>\%
   doubleStack(c('CL', 'V'), c('Cohort', 'Sex')) \%>\%
   ggplot(aes(x = value2, y = value1)) +
   facet_grid(variable1 ~ variable2, scales = 'free') +
   xlab('') + ylab('') +
   geom_boxplot()
}
\seealso{
Other basic: 
\code{\link{asNumeric}()},
\code{\link{compare_classes}()},
\code{\link{doubleStack}()},
\code{\link{doubleStack.default}()},
\code{\link{extract_character}()},
\code{\link{extract_number}()},
\code{\link{fxs}()},
\code{\link{geocv}()},
\code{\link{geomean}()},
\code{\link{isMissing}()},
\code{\link{isNumeric}()},
\code{\link{logit}()},
\code{\link{logit_inv}()},
\code{\link{lunique}()},
\code{\link{make_symmetric}()},
\code{\link{odwg}()},
\code{\link{odwg_classification.data.frame}()},
\code{\link{rgenpois}()},
\code{\link{rnorm_by_id}()},
\code{\link{runc}()},
\code{\link{sample_by_id}()},
\code{\link{sunique}()}
}
\concept{basic}
