% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleStack.r
\name{doubleStack.default}
\alias{doubleStack.default}
\title{Stack a Value Twice by Two Covariates}
\usage{
\method{doubleStack}{default}(x, vars1, vars2, ...)
}
\arguments{
\item{x}{dataset containing all variables specified in vars1 and vars2}

\item{vars1}{The first variable to stack by. For example \code{c(WT,AGE,BMI,SEX)}}

\item{vars2}{The second variable to stack by after the first stack. For example \code{c(ETA1,ETA2,ETA3)}}

\item{...}{ignored}
}
\value{
A data.frame
}
\description{
Dataset will be melted by specified columns after it will be melted by a second set of columns.
This comes in handy when plotting multiple individual parameter estimates versus multiple covariates.
}
\examples{
set.seed(1234)

myDF = data.frame(id = 1:100)
myDF$wt = with(myDF, signif(rnorm_by_id(id, 76,15)))
myDF$age = with(myDF, signif(sample_by_id(id, samples = seq(18,99))))
myDF$sex = with(myDF, sample_by_id(id, samples = c("F","M")))
myDF$CL = with(myDF, signif(rnorm_by_id(id, 10,1)))
myDF$V = with(myDF, signif(rnorm_by_id(id, 3,0.25)))
stacked.df = doubleStack(myDF, vars1 = c("CL","V"), vars2 = c("wt","age","sex"))
stacked.df = doubleStack(myDF, vars1 = c("CL","V"), vars2 = c("wt","age"))
head(myDF)
head(stacked.df)
}
\seealso{
Other basic: 
\code{\link{asNumeric}()},
\code{\link{compare_classes}()},
\code{\link{doubleStack}()},
\code{\link{doubleStack.decorated}()},
\code{\link{extract_character}()},
\code{\link{extract_number}()},
\code{\link{fxs}()},
\code{\link{geocv}()},
\code{\link{geomean}()},
\code{\link{isMissing}()},
\code{\link{isNumeric}()},
\code{\link{logit}()},
\code{\link{logit_inv}()},
\code{\link{lunique}()},
\code{\link{make_symmetric}()},
\code{\link{odwg}()},
\code{\link{odwg_classification.data.frame}()},
\code{\link{rgenpois}()},
\code{\link{rnorm_by_id}()},
\code{\link{runc}()},
\code{\link{sample_by_id}()},
\code{\link{sunique}()}
}
\concept{basic}
