% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_1comp_1abs.r
\name{eff_1comp_1abs}
\alias{eff_1comp_1abs}
\title{Check for missing values}
\usage{
eff_1comp_1abs(dose, tob, parms)
}
\arguments{
\item{dose}{value for dose level}

\item{tob}{numeric vector describing time of observation}

\item{parms}{numeric vector consisting of clearance cl, volume of distribution v, absorption rate constant ka, and effect compartment rate constant keo (in this order). The vector might be named but this is not required.}
}
\value{
A numeric vector of concentration predictions
}
\description{
Create a effect predictions for an effect compartment model based on a 1 compartment open model with first order absorption
}
\examples{
eff_1comp_1abs(dose=1000,tob=seq(0,48,1),parms=c(3,10,0.05))
}
\seealso{
Other modelpred: 
\code{\link{pk_1comp_1abs}()}
}
\concept{modelpred}
