% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_NONMEM_data.r
\name{example_NONMEM_data}
\alias{example_NONMEM_data}
\title{Create an example PKPD Dataset}
\usage{
example_NONMEM_data(ID = 3, TIME = seq(0, 24, 2), DOSE = c(1, 2.5, 10))
}
\arguments{
\item{ID}{vector of ID values}

\item{TIME}{numeric vector describing time of observation}

\item{DOSE}{values for dose level}
}
\value{
A data.frame with concentration and effect predictions
}
\description{
Create a data set with concentration predictions and (effect compartment) predictions for a 1 compartment open model with first order absorption
}
\examples{
pkpd_df = example_NONMEM_data()
head(pkpd_df)
}
\seealso{
Other example: 
\code{\link{example_pkpd_data}()}
}
\concept{example}
