% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_pkpd_data.r
\name{example_pkpd_data}
\alias{example_pkpd_data}
\title{Create an example PKPD Dataset}
\usage{
example_pkpd_data(
  dose = c(0, 100, 250, 500),
  age = c(mid = 50, sd = 10, min = 18, max = 85),
  wt = c(mid = 75, sd = 10, min = 0, max = 140),
  seed = 1234,
  nsub = 32,
  tob = c(0, 1, 2, 3, 4, 5, 6, 7, 10, 14, 21, 28, 35, 42),
  params = c(3, 10, 0.05, 0.05)
)
}
\arguments{
\item{dose}{value for dose level}

\item{age}{numeric vector of length four to populate a normal age distribution with mean (1), SD (2), minimum (3) and maximum (4)}

\item{wt}{numeric vector of length four to populate a normal weight distribution with mean (1), SD (2), minimum (3) and maximum (4)}

\item{seed}{seed for sampling}

\item{nsub}{number of subjects to simulate}

\item{tob}{numeric vector describing time of observation}

\item{params}{numeric vector consisting of clearance, volume of distribution and absorption rate constant (in this order). The vector might be named but this is not required.}
}
\value{
A data.frame with concentration and effect predictions
}
\description{
Create a data set with concentration predictions and (effect compartment) predictions for a 1 compartment open model with first order absorption
}
\examples{
pkpd_df = example_pkpd_data()
head(pkpd_df)
}
\seealso{
Other example: 
\code{\link{example_NONMEM_data}()}
}
\concept{example}
