% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_number.r
\name{extract_number}
\alias{extract_number}
\title{Takes character string and parses the numeric input from it, if available.}
\usage{
extract_number(x)
}
\arguments{
\item{x}{A character or numeric string.}
}
\value{
A numeric vector if the input has multiple character blocks.
}
\description{
Takes character string and parses the numeric input from it, if available.
}
\section{Notes}{

If multiple blocks of numeric input exist wthin the input, only the first is parsed
}

\examples{
extract_number("c76")
extract_number("c76c7")
extract_number("BLOQ = 0.144")
}
\seealso{
Other basic: 
\code{\link{asNumeric}()},
\code{\link{doubleStack}()},
\code{\link{extract_character}()},
\code{\link{fxs}()},
\code{\link{geomean_cv}()},
\code{\link{geomean}()},
\code{\link{isMissing}()},
\code{\link{isNumeric}()},
\code{\link{locf}()},
\code{\link{logit_inv}()},
\code{\link{logit}()},
\code{\link{lunique}()},
\code{\link{make.symmetric}()},
\code{\link{rgenpois}()},
\code{\link{rnorm_by_id}()},
\code{\link{sample_by_id}()},
\code{\link{sunique}()}
}
\concept{basic}
