% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cov.r
\name{extract_varcov}
\alias{extract_varcov}
\title{Get the covariance matrix from an already parsed .cov output file.}
\usage{
extract_varcov(text)
}
\arguments{
\item{text}{A character or numeric string containing the .cov output.}
}
\value{
A named list. The sole item in the list should have a name that is the estimation method
used (e.g. "First Order Conditional Estimation with Interaction").  That named item will be a
data frame containing the covariance matrix.
}
\description{
Get the covariance matrix from an already parsed .cov output file.
}
\section{Notes}{

function not to be called alone - is called by qP function nm.covmat.extract inside get.covmat
}

\examples{
\dontrun{
nm_unzip(
  path = file.path(getOption("qpExampleDir"),"run11"),
  x = "run11", extension = ".cov"
)
x = scan(
  file = file.path(getOption("qpExampleDir"),"run11","run11.cov"),
  what = "character",sep = "\n", quiet = T
)

## the heart of the function: parse space delimited text into a matrix
cov = extract.varcov(x[-1])

## warning: this matrix is LARGE. View only start and end of output
head(cov)
tail(cov)

file.remove(file.path(getOption("qpExampleDir"),"run11","run11.cov")) ## clean things up
}
}
\seealso{
Other nm: 
\code{\link{definitions.nonmem}()},
\code{\link{get_est_table}()},
\code{\link{get_est_table.default}()},
\code{\link{get_est_table.nonmem}()},
\code{\link{get_ofv}()},
\code{\link{get_ofv.character}()},
\code{\link{nm_cov_table}()},
\code{\link{nm_unzip}()},
\code{\link{read_nm_qp}()},
\code{\link{read_nm_qp.character}()},
\code{\link{read_nm_qp.nonmem}()},
\code{\link{sample_uncert_qp}()}
}
\concept{nm}
\keyword{internal}
