% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_diag.r
\name{find_diag.data.frame}
\alias{find_diag.data.frame}
\title{Find Diagonal Elements of a NONMEM Parameter Table for Data Frame}
\usage{
\method{find_diag}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame of class 'ext'}

\item{...}{passed arguments}
}
\value{
data.frame
}
\description{
Find diagonal elements of a NONMEM parameter table for 'data.frame'.
}
\examples{
library(magrittr)
options(nmDir = getOption("qpExampleDir"))
"example2" \%>\% read_ext \%>\% as.list \%>\% summary -> tmp
tmp[[1]] \%>\% find_diag
"example2" \%>\% find_diag

#an error is thrown in case a run does not minimize or terminates without parseable content
#in that case use find_diag_template instead
}
\seealso{
[find_diag_template()]
}
