% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_diag_template.r
\name{find_diag_template}
\alias{find_diag_template}
\title{Create template NULL value Diagonal Elements of a NONMEM Parameter Table}
\usage{
find_diag_template(
  x,
  directory = getOption("nmDir", getwd()),
  extension = c(getOption("nmFileLst", "ext"), getOption("nmFileLst", "lst")),
  quiet = TRUE,
  clean = TRUE,
  nested = c(TRUE, FALSE),
  clue = "$PROB",
  ...
)
}
\arguments{
\item{x}{character: a file path, or the name of a run in 'directory', or ext content}

\item{directory}{character: optional location of x}

\item{extension}{length-two character: extensions identifying ext and lst files}

\item{quiet}{logical: flag for displaying intermediate output}

\item{clean}{should the unzipped file be removed when done (defaults to TRUE)}

\item{nested}{length-two logical: are ext and lst files nested within the run directory?}

\item{...}{passed}
}
\description{
use first component of NONMEM output to read THETA, SIGMA, and OMEGA structure
and turn this into a table that informs which parameters are on-diagonal elements (on=1)
and which are not (on=0). This will be used by \code{\link{nm_params_table}} 
to have a column to (de)select off-diagnal elements in an parameter table.
}
\examples{
library(magrittr)
library(dplyr)
options(nmDir = getOption("qpExampleDir"))
"run105" \%>\% find_diag_template()
}
\seealso{
Other lst: 
\code{\link{coldex}()},
\code{\link{load_modeled_dataset}()},
\code{\link{nm_lst_by_est_method}()},
\code{\link{nm_lst_by_est_method.character}()},
\code{\link{nm_lst_by_est_method.lst}()},
\code{\link{nm_methods}()},
\code{\link{nm_methods.character}()},
\code{\link{nm_methods.lst}()},
\code{\link{nm_methods_short}()},
\code{\link{nm_parse_control_stream}()},
\code{\link{nm_parse_dollar_data}()},
\code{\link{nm_shrinkage}()},
\code{\link{nm_version}()},
\code{\link{nm_version.character}()},
\code{\link{nm_version.lst}()},
\code{\link{read_lst}()},
\code{\link{read_lst.character}()},
\code{\link{read_lst.lst}()},
\code{\link{rowdex}()}
}
\concept{lst}
\keyword{internal}
