% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_params_table.r
\name{format.nmpartab_}
\alias{format.nmpartab_}
\title{Format Parameter Table Element}
\usage{
\method{format}{nmpartab_}(
  x,
  align.dot = TRUE,
  latex = FALSE,
  digits = 3,
  fixed.text = "(fixed)",
  paste_label_unit = "none",
  ci.sep = " - ",
  backtransform = TRUE,
  pretty = TRUE,
  ...
)
}
\arguments{
\item{x}{data.frame}

\item{align.dot}{logical: align dots around decimal point, defaults to TRUE, but only effective when latex=TRUE}

\item{latex}{logical: creates escaped \LaTeX output and effectuates align.dot argument. Defaults to FALSE.}

\item{digits}{numeric: number of significant digits assed to (formatted_)signif. Defaults to 3.}

\item{fixed.text}{character: string used as suffix after a fixed parameter estimate.}

\item{paste_label_unit}{character: Must be either "none" (default) keeping label and unit separate columns or "label" which glue label and bracketed unit together. Defaults to "none"}

\item{ci.sep}{character: character to separate confidence intervals by; only updates where transform is LOG or LOGIT; coordinate with usage in \code{\link{nm_params_table}}}

\item{backtransform}{whether to backtransform estimate, CI95, and RSE (LOG, LOGIT)}

\item{pretty}{provide the key names after formatting only (TRUE, default) or keep the untransformed columns as well (FALSE)}

\item{...}{currently unused arguments}
}
\value{
data.frame; column 'Name' has class 'nms_nonmem'
}
\description{
Formats parameter tables for inclusion in reports, typically in \LaTeX or Rmarkdown 
or bookdown format. 
It takes one element of class 'nmpartab' which \code{\link{nm_params_table}} and 
\code{\link{nm_params_table_short}} provide. 
Alternative methods for parameter tables: see also \code{\link{get_est_table}} 
which uses \pkg{pmxTools} to construct a parameter table from the XML. 
See also \code{\link[nonmemica]{partab}} from \pkg{nonmemica} which does similar.
Formatting functions \code{\link{formatted_signif}} that calls 
\code{\link{align_around_decimal_point}}.
}
\seealso{
Other ext: 
\code{\link{as.list.ext}()},
\code{\link{as.matrix.ext}()},
\code{\link{find_diag}()},
\code{\link{format.nmpartab}()},
\code{\link{get_ext}()},
\code{\link{get_ext.character}()},
\code{\link{get_ext.ext}()},
\code{\link{get_ext.extList}()},
\code{\link{get_ofv.ext}()},
\code{\link{get_ofv.nonmem}()},
\code{\link{lol}()},
\code{\link{nm_condition_number}()},
\code{\link{nm_condition_number.default}()},
\code{\link{nm_condition_number.ext}()},
\code{\link{nm_params_table}()},
\code{\link{nm_params_table.character}()},
\code{\link{nm_params_table_short}()},
\code{\link{nm_params_table_short.character}()},
\code{\link{nm_params_table_short.ext}()},
\code{\link{nm_params_table_short.nmpartab}()},
\code{\link{read_ext}()},
\code{\link{read_ext.character}()},
\code{\link{read_ext.ext}()},
\code{\link{summary.extList}()}
}
\concept{ext}
\keyword{internal}
