% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxs.r
\name{fxs}
\alias{fxs}
\title{Fixed character length of a String}
\usage{
fxs(x, pattern = "0000")
}
\arguments{
\item{x}{string to format}

\item{pattern}{format to adhere to}
}
\value{
x reformatted according to pattern
}
\description{
Format string of fixed character length
}
\examples{
fxs(c(0,1,34, 50000), pattern = "    ")
fxs(c(0,1,34, 50000), "000000")
}
\seealso{
Other basic: 
\code{\link{asNumeric}()},
\code{\link{doubleStack}()},
\code{\link{extract_character}()},
\code{\link{extract_number}()},
\code{\link{geomean_cv}()},
\code{\link{geomean}()},
\code{\link{isMissing}()},
\code{\link{isNumeric}()},
\code{\link{locf}()},
\code{\link{logit_inv}()},
\code{\link{logit}()},
\code{\link{lunique}()},
\code{\link{make.symmetric}()},
\code{\link{rgenpois}()},
\code{\link{rnorm_by_id}()},
\code{\link{sample_by_id}()},
\code{\link{sunique}()}
}
\concept{basic}
