% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocv.r
\name{geocv}
\alias{geocv}
\title{Geometric Coefficient of Variation}
\usage{
geocv(x, na.rm = FALSE)
}
\arguments{
\item{x}{presumed log-normally distributed values}

\item{na.rm}{logical: set TRUE to omit NA entries}
}
\value{
geometric coefficient of variation (not percent)
}
\description{
Geometric Coefficient of Variation
}
\examples{
set.seed(1234)
x = rlnorm(1000)
mean(x)
geomean(x)
geocv(x)
}
\seealso{
Other basic: 
\code{\link{asNumeric}()},
\code{\link{compare_classes}()},
\code{\link{doubleStack}()},
\code{\link{doubleStack.decorated}()},
\code{\link{doubleStack.default}()},
\code{\link{extract_character}()},
\code{\link{extract_number}()},
\code{\link{fxs}()},
\code{\link{geomean}()},
\code{\link{isMissing}()},
\code{\link{isNumeric}()},
\code{\link{logit}()},
\code{\link{logit_inv}()},
\code{\link{lunique}()},
\code{\link{make_symmetric}()},
\code{\link{odwg}()},
\code{\link{odwg_classification.data.frame}()},
\code{\link{rgenpois}()},
\code{\link{rnorm_by_id}()},
\code{\link{runc}()},
\code{\link{sample_by_id}()},
\code{\link{sunique}()}
}
\concept{basic}
