% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomean.r
\name{geomean}
\alias{geomean}
\title{Geometric Mean}
\usage{
geomean(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric value}

\item{na.rm}{logical (default to TRUE) to omit NA entries}
}
\value{
Geometric mean of vector x
}
\description{
Geometric Mean
}
\examples{
set.seed(1234)
x = rlnorm(1000)
mean(x)
geomean(x)
}
\seealso{
Other basic: 
\code{\link{asNumeric}()},
\code{\link{doubleStack}()},
\code{\link{extract_character}()},
\code{\link{extract_number}()},
\code{\link{fxs}()},
\code{\link{geomean_cv}()},
\code{\link{isMissing}()},
\code{\link{isNumeric}()},
\code{\link{locf}()},
\code{\link{logit_inv}()},
\code{\link{logit}()},
\code{\link{lunique}()},
\code{\link{make.symmetric}()},
\code{\link{rgenpois}()},
\code{\link{rnorm_by_id}()},
\code{\link{sample_by_id}()},
\code{\link{sunique}()}
}
\concept{basic}
