% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomean_cv.r
\name{geomean_cv}
\alias{geomean_cv}
\title{Geometric Mean Coefficient of Variation in Percent}
\usage{
geomean_cv(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric values}

\item{na.rm}{logical (default to TRUE) to omit NA entries}
}
\value{
Geometric mean of vector x
}
\description{
Geometric Mean Coefficient of Variation in Percent
}
\examples{
set.seed(1234)
x = rlnorm(1000)
mean(x)
geomean(x)
geomean_cv(x)
}
\seealso{
Other basic: 
\code{\link{asNumeric}()},
\code{\link{doubleStack}()},
\code{\link{extract_character}()},
\code{\link{extract_number}()},
\code{\link{fxs}()},
\code{\link{geomean}()},
\code{\link{isMissing}()},
\code{\link{isNumeric}()},
\code{\link{locf}()},
\code{\link{logit_inv}()},
\code{\link{logit}()},
\code{\link{lunique}()},
\code{\link{make.symmetric}()},
\code{\link{rgenpois}()},
\code{\link{rnorm_by_id}()},
\code{\link{sample_by_id}()},
\code{\link{sunique}()}
}
\concept{basic}
