% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_coi.r
\name{get_coimat.character}
\alias{get_coimat.character}
\title{Get Inverse Correlation Matrix for Character}
\usage{
\method{get_coimat}{character}(
  x,
  directory = getOption("nmDir", getwd()),
  extension = getOption("nmFilecoi", "coi"),
  quiet = TRUE,
  clean = TRUE,
  nested = TRUE,
  clue = "TABLE NO.",
  ...
)
}
\arguments{
\item{x}{character: a file path, or the name of a run in 'directory', or coi content}

\item{directory}{character: optional location of x}

\item{extension}{character: extension identifying coi files}

\item{quiet}{logical: flag for displaying intermediate output}

\item{clean}{should the unzipped file be removed when done (defaults to TRUE)}

\item{nested}{is the file nested within the run directory?}

\item{clue}{a text fragment expected to occur in coi files}

\item{...}{passed arguments}
}
\value{
list of Inverse Correlation matrices as data frames
}
\description{
Gets Inverse Correlation matrix for character.
}
\seealso{
Other coi: 
\code{\link{get_coimat}()},
\code{\link{get_coimat.coi}()},
\code{\link{read_coi}()},
\code{\link{read_coi.character}()},
\code{\link{read_coi.coi}()}
}
\concept{coi}
