% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_coi.r
\name{get_coimat.coi}
\alias{get_coimat.coi}
\title{Get Inverse Correlation Matrix for COI}
\usage{
\method{get_coimat}{coi}(x, ...)
}
\arguments{
\item{x}{coi}

\item{...}{passed arguments}
}
\value{
list of Inverse Correlation matrices as data frames
}
\description{
Gets Inverse Correlation matrix for COI.
}
\seealso{
Other coi: 
\code{\link{get_coimat}()},
\code{\link{get_coimat.character}()},
\code{\link{read_coi}()},
\code{\link{read_coi.character}()},
\code{\link{read_coi.coi}()}
}
\concept{coi}
\keyword{internal}
