% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_coi.r
\name{get_coimat}
\alias{get_coimat}
\title{Get Inverse Correlation Matrix}
\usage{
get_coimat(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Gets Inverse Inverse Correlation matrix.
Generic, with methods \code{\link{get_coimat.character}} and \code{\link{get_coimat.coi}}.
}
\seealso{
Other coi: 
\code{\link{get_coimat.character}()},
\code{\link{get_coimat.coi}()},
\code{\link{read_coi}()},
\code{\link{read_coi.character}()},
\code{\link{read_coi.coi}()}
}
\concept{coi}
\keyword{internal}
