% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cor.r
\name{get_cormat.character}
\alias{get_cormat.character}
\title{Get Correlation Matrix for Character}
\usage{
\method{get_cormat}{character}(
  x,
  directory = getOption("nmDir", getwd()),
  extension = getOption("nmFilecor", "cor"),
  quiet = TRUE,
  clean = TRUE,
  nested = TRUE,
  clue = "TABLE NO.",
  ...
)
}
\arguments{
\item{x}{character: a file path, or the name of a run in 'directory', or cor content}

\item{directory}{character: optional location of x}

\item{extension}{character: extension identifying cor files}

\item{quiet}{logical: flag for displaying intermediate output}

\item{clean}{should the unzipped file be removed when done (defaults to TRUE)}

\item{nested}{is the file nested within the run directory?}

\item{clue}{a text fragment expected to occur in cor files}

\item{...}{passed arguments}
}
\value{
list of correlation matrices as data frames
}
\description{
Gets Correlation matrix for character.
}
\seealso{
Other cor: 
\code{\link{get_cormat}()},
\code{\link{get_cormat.cor}()},
\code{\link{read_cor}()},
\code{\link{read_cor.character}()},
\code{\link{read_cor.cor}()}
}
\concept{cor}
