% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cor.r
\name{get_cormat.cor}
\alias{get_cormat.cor}
\title{Get Correlation Matrix for COR}
\usage{
\method{get_cormat}{cor}(x, ...)
}
\arguments{
\item{x}{cor}

\item{...}{passed arguments}
}
\value{
list of correlation matrices as data frames
}
\description{
Gets correlation matrix for COR.
}
\seealso{
Other cor: 
\code{\link{get_cormat}()},
\code{\link{get_cormat.character}()},
\code{\link{read_cor}()},
\code{\link{read_cor.character}()},
\code{\link{read_cor.cor}()}
}
\concept{cor}
\keyword{internal}
