% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cor.r
\name{get_cormat}
\alias{get_cormat}
\title{Get Correlation Matrix}
\usage{
get_cormat(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Gets correlation matrix.
Generic, with methods \code{\link{get_cormat.character}} and \code{\link{get_cormat.cor}}.
}
\seealso{
Other cor: 
\code{\link{get_cormat.character}()},
\code{\link{get_cormat.cor}()},
\code{\link{read_cor}()},
\code{\link{read_cor.character}()},
\code{\link{read_cor.cor}()}
}
\concept{cor}
\keyword{internal}
