% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cov.r
\name{get_covmat.cov}
\alias{get_covmat.cov}
\title{Get Covariance Matrix for COV}
\usage{
\method{get_covmat}{cov}(x, ...)
}
\arguments{
\item{x}{cov}

\item{...}{passed arguments}
}
\value{
list of covariance matrices as data frames
}
\description{
Get covariance matrix for COV.
}
\seealso{
Other cov: 
\code{\link{get_covmat}()},
\code{\link{get_covmat.character}()},
\code{\link{read_cov}()},
\code{\link{read_cov.character}()},
\code{\link{read_cov.cov}()}
}
\concept{cov}
\keyword{internal}
