% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cov.r
\name{get_covmat}
\alias{get_covmat}
\title{Get Covariance Matrix}
\usage{
get_covmat(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Gets covariance matrix.
Generic, with methods \code{\link{get_covmat.character}} and \code{\link{get_covmat.cov}}.
}
\seealso{
Other cov: 
\code{\link{get_covmat.character}()},
\code{\link{get_covmat.cov}()},
\code{\link{read_cov}()},
\code{\link{read_cov.character}()},
\code{\link{read_cov.cov}()}
}
\concept{cov}
\keyword{internal}
