% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm.r
\name{get_est_table.default}
\alias{get_est_table.default}
\title{Create a Table of Model Parameters by Default}
\usage{
\method{get_est_table}{default}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed to \code{\link[pmxTools]{get_est_table}}}
}
\description{
Creates a Table of Model Parameters using default method.
See also \code{\link[pmxTools]{get_est_table}}.
}
\seealso{
Other nm: 
\code{\link{as_pmx_canonical}()},
\code{\link{definitions.nonmem}()},
\code{\link{extract_varcov}()},
\code{\link{get_est_table.nonmem}()},
\code{\link{get_est_table}()},
\code{\link{nm_cov_table}()},
\code{\link{nm_covmat_extract}()},
\code{\link{nm_unzip}()},
\code{\link{read_nm.character}()},
\code{\link{read_nm.file}()},
\code{\link{read_nm.run}()},
\code{\link{read_nm.xml}()},
\code{\link{read_nm}()},
\code{\link{sample_uncert}()}
}
\concept{nm}
