% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ext.r
\name{get_ext.character}
\alias{get_ext.character}
\title{Get Ext Summary for Character}
\usage{
\method{get_ext}{character}(
  x,
  directory = getOption("nmDir", getwd()),
  extension = getOption("nmFileExt", "ext"),
  quiet = TRUE,
  clean = TRUE,
  nested = TRUE,
  clue = "ITERATION",
  final = FALSE,
  ...
)
}
\arguments{
\item{x}{character: a file path, or the name of a run in 'directory', or cov content}

\item{directory}{character: optional location of x}

\item{extension}{character: extension identifying cov files}

\item{quiet}{logical: flag for displaying intermediate output}

\item{clean}{should the unzipped file be removed when done (defaults to TRUE)}

\item{nested}{is the file nested within the run directory?}

\item{clue}{a text fragment expected to occur in cov files}

\item{final}{logical: return just the iterations with numbers greater than zero}

\item{...}{passed arguments}
}
\value{
list of covariance matrices as data frames
}
\description{
Get ext summary for character.
}
\seealso{
Other ext: 
\code{\link{as.list.ext}()},
\code{\link{as.matrix.ext}()},
\code{\link{find_diag}()},
\code{\link{format.nmpartab}()},
\code{\link{format.nmpartab_}()},
\code{\link{get_ext}()},
\code{\link{get_ext.ext}()},
\code{\link{get_ext.extList}()},
\code{\link{get_ofv.ext}()},
\code{\link{get_ofv.nonmem}()},
\code{\link{lol}()},
\code{\link{nm_condition_number}()},
\code{\link{nm_condition_number.default}()},
\code{\link{nm_condition_number.ext}()},
\code{\link{nm_params_table}()},
\code{\link{nm_params_table.character}()},
\code{\link{nm_params_table_short}()},
\code{\link{nm_params_table_short.character}()},
\code{\link{nm_params_table_short.ext}()},
\code{\link{nm_params_table_short.nmpartab}()},
\code{\link{read_ext}()},
\code{\link{read_ext.character}()},
\code{\link{read_ext.ext}()},
\code{\link{summary.extList}()}
}
\concept{ext}
