% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ext.r
\name{get_ext.ext}
\alias{get_ext.ext}
\title{Get EXT Summary for EXT}
\usage{
\method{get_ext}{ext}(x, final = FALSE, ...)
}
\arguments{
\item{x}{ext}

\item{final}{logical: return just the iterations with numbers greater than zero}

\item{...}{passed arguments}
}
\value{
list of covariance matrices as data frames
}
\description{
Get EXT Summary for EXT
}
\seealso{
Other ext: 
\code{\link{as.list.ext}()},
\code{\link{as.matrix.ext}()},
\code{\link{find_diag}()},
\code{\link{format.nmpartab}()},
\code{\link{format.nmpartab_}()},
\code{\link{get_ext}()},
\code{\link{get_ext.character}()},
\code{\link{get_ext.extList}()},
\code{\link{get_ofv.ext}()},
\code{\link{get_ofv.nonmem}()},
\code{\link{lol}()},
\code{\link{nm_condition_number}()},
\code{\link{nm_condition_number.default}()},
\code{\link{nm_condition_number.ext}()},
\code{\link{nm_params_table}()},
\code{\link{nm_params_table.character}()},
\code{\link{nm_params_table_short}()},
\code{\link{nm_params_table_short.character}()},
\code{\link{nm_params_table_short.ext}()},
\code{\link{nm_params_table_short.nmpartab}()},
\code{\link{read_ext}()},
\code{\link{read_ext.character}()},
\code{\link{read_ext.ext}()},
\code{\link{summary.extList}()}
}
\concept{ext}
\keyword{internal}
