% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ofv.r
\name{get_ofv.character}
\alias{get_ofv.character}
\title{Retrieve NONMEM Objective Function Value for Character}
\usage{
\method{get_ofv}{character}(
  x,
  directory = getOption("nmDir", getwd()),
  quiet = TRUE,
  clean = TRUE,
  digits = 1,
  ...
)
}
\arguments{
\item{x}{character: a file path, or the name of a run in 'directory', or xml content}

\item{directory}{character: optional location of x}

\item{quiet}{logical: flag for displaying intermediate output}

\item{clean}{should the unzipped file be removed when done (defaults to TRUE)}

\item{digits}{passed to \code{\link{round}}}

\item{...}{passed arguments}
}
\value{
numeric
}
\description{
Retrieves NONMEM Objective Function Value(s) for class character.
An attempt is made to acquire the xml output for the relevant model run.
Then nodes matching "final_objective_function" are extracted, rounded, 
and returned.
}
\examples{
library(magrittr)
dir <- getOption("qpExampleDir")
options(nmDir = dir)
run <- 'example1'
file <- 'example1.ext'
get_ofv(run)
get_ofv(run, directory = dir)
get_ofv(file)
get_ofv(file, directory = dir)
}
\seealso{
Other nm: 
\code{\link{definitions.nonmem}()},
\code{\link{extract_varcov}()},
\code{\link{get_est_table}()},
\code{\link{get_est_table.default}()},
\code{\link{get_est_table.nonmem}()},
\code{\link{get_ofv}()},
\code{\link{nm_cov_table}()},
\code{\link{nm_unzip}()},
\code{\link{read_nm_qp}()},
\code{\link{read_nm_qp.character}()},
\code{\link{read_nm_qp.nonmem}()},
\code{\link{sample_uncert_qp}()}
}
\concept{nm}
