% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ofv.r
\name{get_ofv.ext}
\alias{get_ofv.ext}
\title{Retrieve NONMEM Objective Function Value for Ext}
\usage{
\method{get_ofv}{ext}(x, digits = 1, ...)
}
\arguments{
\item{x}{character: a file path, or the name of a run in 'directory', or xml content}

\item{digits}{passed to \code{\link{round}}}

\item{...}{ignored}
}
\value{
numeric
}
\description{
Retrieves NONMEM Objective Function Value(s) for class 'ext'. 
See \code{\link{read_ext.character}} and \code{\link{get_ofv.character}}.
}
\seealso{
Other ext: 
\code{\link{as.list.ext}()},
\code{\link{as.matrix.ext}()},
\code{\link{find_diag}()},
\code{\link{format.nmpartab}()},
\code{\link{format.nmpartab_}()},
\code{\link{get_ext}()},
\code{\link{get_ext.character}()},
\code{\link{get_ext.ext}()},
\code{\link{get_ext.extList}()},
\code{\link{get_ofv.nonmem}()},
\code{\link{lol}()},
\code{\link{nm_condition_number}()},
\code{\link{nm_condition_number.default}()},
\code{\link{nm_condition_number.ext}()},
\code{\link{nm_params_table}()},
\code{\link{nm_params_table.character}()},
\code{\link{nm_params_table_short}()},
\code{\link{nm_params_table_short.character}()},
\code{\link{nm_params_table_short.ext}()},
\code{\link{nm_params_table_short.nmpartab}()},
\code{\link{read_ext}()},
\code{\link{read_ext.character}()},
\code{\link{read_ext.ext}()},
\code{\link{summary.extList}()}
}
\concept{ext}
