% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ofv.r
\name{get_ofv}
\alias{get_ofv}
\title{Retrieve NONMEM Objective Function Value}
\usage{
get_ofv(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed}
}
\description{
Retrieves NONMEM Objective Function Value(s).
Generic, with method \code{\link{get_ofv.character}}
}
\seealso{
Other nm: 
\code{\link{definitions.nonmem}()},
\code{\link{extract_varcov}()},
\code{\link{get_est_table}()},
\code{\link{get_est_table.default}()},
\code{\link{get_est_table.nonmem}()},
\code{\link{get_ofv.character}()},
\code{\link{nm_cov_table}()},
\code{\link{nm_unzip}()},
\code{\link{read_nm_qp}()},
\code{\link{read_nm_qp.character}()},
\code{\link{read_nm_qp.nonmem}()},
\code{\link{sample_uncert_qp}()}
}
\concept{nm}
\keyword{internal}
