% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xpose_tables.r
\name{get_xpose_tables}
\alias{get_xpose_tables}
\title{Compile all Xpose Output Datasets into one Dataset}
\usage{
get_xpose_tables(
  x,
  directory = getOption("nmDir"),
  pattern = "(sd|pa|ca|co|my|sim)tab|.tab",
  ...
)
}
\arguments{
\item{x}{character vector of run names, e.g. c('run1','run2')}

\item{directory}{path to working directory}

\item{pattern}{regular expression matching xpose table file names}

\item{...}{ignored}
}
\value{
data.frame
}
\description{
Compiles a NONMEM output files as single dataset.
}
\examples{
library(magrittr)
options(nmDir = getOption('qpExampleDir'))
'run100' \%>\% get_xpose_tables
}
\seealso{
Other runrecord: 
\code{\link{nm_runrecord}()},
\code{\link{nm_runrecord_graph}()}
}
\concept{runrecord}
