% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isMissing.r
\name{isMissing}
\alias{isMissing}
\title{Check for missing values}
\usage{
isMissing(x)
}
\arguments{
\item{x}{vector}
}
\value{
A vector of boolean values indicating which x is missing
}
\description{
Take a vector and return a vector of boolean values indicating which input vector values are missing (NA).
}
\examples{
temp = data.frame(id = 1:4, nom.time = c(5,NA,7,NA))
isMissing(temp$nom.time)
temp[isMissing(temp$nom.time),]
}
\seealso{
Other basic: 
\code{\link{asNumeric}()},
\code{\link{doubleStack}()},
\code{\link{extract_character}()},
\code{\link{extract_number}()},
\code{\link{fxs}()},
\code{\link{geomean_cv}()},
\code{\link{geomean}()},
\code{\link{isNumeric}()},
\code{\link{locf}()},
\code{\link{logit_inv}()},
\code{\link{logit}()},
\code{\link{lunique}()},
\code{\link{make.symmetric}()},
\code{\link{rgenpois}()},
\code{\link{rnorm_by_id}()},
\code{\link{sample_by_id}()},
\code{\link{sunique}()}
}
\concept{basic}
