% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isNumeric.r
\name{isNumeric}
\alias{isNumeric}
\title{Check for numeric values}
\usage{
isNumeric(x)
}
\arguments{
\item{x}{Placehoder for parameter description}
}
\value{
Placeholder for return description
}
\description{
Take a vector and return a vector of boolean values indicating which input
}
\examples{
isNumeric(12)
isNumeric("x")
isNumeric(c(1,2,3,"b",5,6))
}
\seealso{
Other basic: 
\code{\link{asNumeric}()},
\code{\link{doubleStack}()},
\code{\link{extract_character}()},
\code{\link{extract_number}()},
\code{\link{fxs}()},
\code{\link{geomean_cv}()},
\code{\link{geomean}()},
\code{\link{isMissing}()},
\code{\link{locf}()},
\code{\link{logit_inv}()},
\code{\link{logit}()},
\code{\link{lunique}()},
\code{\link{make.symmetric}()},
\code{\link{rgenpois}()},
\code{\link{rnorm_by_id}()},
\code{\link{sample_by_id}()},
\code{\link{sunique}()}
}
\concept{basic}
