% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.r
\name{isopair}
\alias{isopair}
\title{Plot Two Isoplots}
\usage{
isopair(data, ...)
}
\arguments{
\item{data}{object of dispatch}

\item{...}{passed}
}
\description{
Plots two isoplots.
Generic, with method \code{\link{isopair.data.frame}}.
}
\seealso{
Other plots: 
\code{\link{isopair.data.frame}()},
\code{\link{isoplot}()},
\code{\link{isoplot.data.frame}()},
\code{\link{trendpair}()},
\code{\link{trendpair.data.frame}()},
\code{\link{trendplot}()},
\code{\link{trendplot.data.frame}()}
}
\concept{plots}
\keyword{internal}
