% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.r
\name{isoplot.data.frame}
\alias{isoplot.data.frame}
\title{Plot Dataframe Isometrically}
\usage{
\method{isoplot}{data.frame}(data, x, y, ..., alpha = 0.5, transform = "identity")
}
\arguments{
\item{data}{data.frame}

\item{x}{bareword for column mapped to x axis}

\item{y}{bareword for column mapped to y axis}

\item{...}{ignored}

\item{alpha}{passed to \code{\link[ggplot2]{geom_point}}}

\item{transform}{passed to \code{\link[ggplot2]{scale_x_continuous}} and \code{\link[ggplot2]{scale_y_continuous}}}
}
\value{
ggplot
}
\description{
Plots 'data.frame' isometrically.
Constrains axes to be identical;
adds reference and trend lines.
Exposes alpha and transformation.
}
\examples{
library(magrittr)
library(yamlet)
x <- 'extdata/modeling/xanomeline-mod.csv.gz' \%>\% 
system.file(package = 'yamlet') \%>\% io_csv \%>\% filter(cp > 1) \%>\% resolve
x \%>\% isoplot(PRED, DV, transform = 'log10')
}
\seealso{
Other plots: 
\code{\link{isopair}()},
\code{\link{isopair.data.frame}()},
\code{\link{isoplot}()},
\code{\link{trendpair}()},
\code{\link{trendpair.data.frame}()},
\code{\link{trendplot}()},
\code{\link{trendplot.data.frame}()}
}
\concept{plots}
\keyword{visualization}
