% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.r
\name{isoplot}
\alias{isoplot}
\title{Plot Something Isometrically}
\usage{
isoplot(data, ...)
}
\arguments{
\item{data}{object of dispatch}

\item{...}{passed}
}
\description{
Plots something isometrically.
Generic, with method \code{\link{isoplot.data.frame}}.
}
\seealso{
Other plots: 
\code{\link{isopair}()},
\code{\link{isopair.data.frame}()},
\code{\link{isoplot.data.frame}()},
\code{\link{trendpair}()},
\code{\link{trendpair.data.frame}()},
\code{\link{trendplot}()},
\code{\link{trendplot.data.frame}()}
}
\concept{plots}
\keyword{internal}
