% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_modeled_dataset.r
\name{load_modeled_dataset}
\alias{load_modeled_dataset}
\title{Load Modeled Dataset}
\usage{
load_modeled_dataset(
  x,
  directory = getOption("nmDir", getwd()),
  extension = getOption("nmFileMod", "mod"),
  visibility = FALSE,
  read.input = list(read.csv, header = TRUE, as.is = TRUE),
  nested = getOption("nested", psn_nested),
  ...
)
}
\arguments{
\item{x}{character run name}

\item{directory}{path to directory containing the run}

\item{extension}{extension for model file, e.g. 'mod' or 'ctl'}

\item{visibility}{logical: include VISIBLE flag indicating ignore/accept results?}

\item{read.input}{if visibility is TRUE, passed to \code{\link[nonmemica]{ignored}}}

\item{nested}{passed to \code{\link[nonmemica]{datafile}} and \code{\link[nonmemica]{ignored}}}

\item{...}{passed to 
\code{\link[nonmemica]{ignored}},
\code{\link[nonmemica]{datafile}}, and
\code{\link[yamlet]{io_csv}}}
}
\value{
data.frame
}
\description{
Returns the modeled dataset, given details about the model run.
If visibility is TRUE, the column VISIBLE is added (it is an error if 
VISIBLE already present) with integer 1 or 0 indicating which records
were visible to NONMEM.
}
\examples{
library(magrittr)
dir <- getOption('qpExampleDir')
options(nmDir = dir)

# This example breaks because we need to skip the first line of the dataset:

'example2' \%>\%
  load_modeled_dataset(extension = 'ctl', nested = FALSE) \%>\% 
  head(3)

# This example is difficult but correct:

load_modeled_dataset(
  'example2',              # exists in the example NONMEM directory
  directory = dir,         # created above
  extension = 'ctl',        # NOT mod, creates problems below
  nested = FALSE,          # cannot guess whether ctl files are inside run directories
  read.input = list(       # a method to read the data for visibility check
    read.csv, 
    header = TRUE, 
    skip = 1               # associated dataset has intro comment! passed to read.csv
  ),
  visibility = TRUE,       # call for visibility check
  skip = 1                 # passed again to io_csv() when acquiring data!
) \%>\% head(3)
}
\seealso{
Other lst: 
\code{\link{coldex}()},
\code{\link{find_diag_template}()},
\code{\link{nm_lst_by_est_method}()},
\code{\link{nm_lst_by_est_method.character}()},
\code{\link{nm_lst_by_est_method.lst}()},
\code{\link{nm_methods}()},
\code{\link{nm_methods.character}()},
\code{\link{nm_methods.lst}()},
\code{\link{nm_methods_short}()},
\code{\link{nm_parse_control_stream}()},
\code{\link{nm_parse_dollar_data}()},
\code{\link{nm_shrinkage}()},
\code{\link{nm_version}()},
\code{\link{nm_version.character}()},
\code{\link{nm_version.lst}()},
\code{\link{read_lst}()},
\code{\link{read_lst.character}()},
\code{\link{read_lst.lst}()},
\code{\link{rowdex}()}
}
\concept{lst}
