% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.r
\name{logit}
\alias{logit}
\title{Logit transformation}
\usage{
logit(x, min = 0, max = 1)
}
\arguments{
\item{x}{any value}

\item{min}{lower boundary of backtransformed value}

\item{max}{upper boundary of backtransformed value}
}
\value{
Logit-transformed x
}
\description{
Logit-transforms any value between 0 and 1
}
\examples{
myVector = runif(n = 1000, min = 0.001,max = 0.999)
myVector.logit = logit(myVector)
par(mfrow = c(1,3))
hist(myVector)
hist(myVector.logit)
hist(logit_inv(myVector.logit))
}
\seealso{
Other basic: 
\code{\link{asNumeric}()},
\code{\link{compare_classes}()},
\code{\link{doubleStack}()},
\code{\link{doubleStack.decorated}()},
\code{\link{doubleStack.default}()},
\code{\link{extract_character}()},
\code{\link{extract_number}()},
\code{\link{fxs}()},
\code{\link{geocv}()},
\code{\link{geomean}()},
\code{\link{isMissing}()},
\code{\link{isNumeric}()},
\code{\link{logit_inv}()},
\code{\link{lunique}()},
\code{\link{make_symmetric}()},
\code{\link{odwg}()},
\code{\link{odwg_classification.data.frame}()},
\code{\link{rgenpois}()},
\code{\link{rnorm_by_id}()},
\code{\link{runc}()},
\code{\link{sample_by_id}()},
\code{\link{sunique}()}
}
\concept{basic}
