% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_inv.r
\name{logit_inv}
\alias{logit_inv}
\title{Logit Back-transformation}
\usage{
logit_inv(x, min = 0, max = 1)
}
\arguments{
\item{x}{any value}

\item{min}{lower boundary of backtransformed value}

\item{max}{upper boundary of backtransformed value}
}
\value{
logit(x) backtransformed to <0,1>
}
\description{
Backtransforms logit transformed data to the <0,1> scale
}
\examples{
myVector = runif(n = 1000, min = 0.001,max = 0.999)
myVector.logit = logit(myVector)
par(mfrow = c(1,3))
hist(myVector)
hist(myVector.logit)
hist(logit_inv(myVector.logit))
}
\seealso{
Other basic: 
\code{\link{asNumeric}()},
\code{\link{doubleStack}()},
\code{\link{extract_character}()},
\code{\link{extract_number}()},
\code{\link{fxs}()},
\code{\link{geomean_cv}()},
\code{\link{geomean}()},
\code{\link{isMissing}()},
\code{\link{isNumeric}()},
\code{\link{locf}()},
\code{\link{logit}()},
\code{\link{lunique}()},
\code{\link{make.symmetric}()},
\code{\link{rgenpois}()},
\code{\link{rnorm_by_id}()},
\code{\link{sample_by_id}()},
\code{\link{sunique}()}
}
\concept{basic}
