% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_symmetric.r
\name{make_symmetric}
\alias{make_symmetric}
\title{Create symmetrical matrix}
\usage{
make_symmetric(values, triangle = "upper")
}
\arguments{
\item{values}{vector of length cumsum(ncol(matrix))}

\item{triangle}{defaulting to 'upper', can also take the value 'lower'}
}
\value{
A matrix of dimension ncol x ncol
}
\description{
Create symmetrical matrix of lower or upper traingle of a matrix
}
\examples{
  make_symmetric(c(0.12,
                   0.02,1.02,
                   0.23,-0.3,0.23))
}
\seealso{
Other basic: 
\code{\link{asNumeric}()},
\code{\link{compare_classes}()},
\code{\link{doubleStack}()},
\code{\link{doubleStack.decorated}()},
\code{\link{doubleStack.default}()},
\code{\link{extract_character}()},
\code{\link{extract_number}()},
\code{\link{fxs}()},
\code{\link{geocv}()},
\code{\link{geomean}()},
\code{\link{isMissing}()},
\code{\link{isNumeric}()},
\code{\link{logit}()},
\code{\link{logit_inv}()},
\code{\link{lunique}()},
\code{\link{odwg}()},
\code{\link{odwg_classification.data.frame}()},
\code{\link{rgenpois}()},
\code{\link{rnorm_by_id}()},
\code{\link{runc}()},
\code{\link{sample_by_id}()},
\code{\link{sunique}()}
}
\concept{basic}
