% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_cov_table.r
\name{nm_cov_table}
\alias{nm_cov_table}
\title{Double Stacking or Wide to Tall Transformation of Parameter and ETA Estimates with Covariates}
\usage{
nm_cov_table(
  x,
  ID = "ID",
  parameters = c("CL", "V", "KA", "KTR"),
  covariates = c("DOSE", "WT", "HT", "AGE", "BMI", "SEX", "RACE"),
  carryAlong = "SUBJID",
  wide2tall = "none",
  n_cat = 10
)
}
\arguments{
\item{x}{NONMEM output tables as a data.frame from get.xpose.tables, or read_nmtables, or alike}

\item{ID}{the subject identifier in the NONMEM output}

\item{parameters}{Model parameter name vector}

\item{covariates}{covariates to integrate}

\item{carryAlong}{columns to carry along i the output}

\item{wide2tall}{either 'parameter' or 'eta'. Plural will be accepted too}

\item{n_cat}{number of unique values to qualify as a categorical covariate}
}
\value{
data frame with parValue, etaValue, covValue, and cov_class (continuous or categorical)
}
\description{
Double stack parameter estimates and ETA values with covariates into a single data frame
}
\examples{

print("hello world")

}
\seealso{
Other nm: 
\code{\link{extract_varcov}()},
\code{\link{nm_covmat_extract}()},
\code{\link{nm_unzip}()},
\code{\link{read_nm_qp}()},
\code{\link{sample_uncert_qp}()}
}
\concept{nm}
