% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_covmat_extract.r
\name{nm_covmat_extract}
\alias{nm_covmat_extract}
\title{Extract the Variance-Covariance Matrix of a Converged NONMEM run}
\usage{
nm_covmat_extract(
  run,
  path = getOption("nmDir"),
  extension = "cov",
  clear_zip = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{run}{character string in the form of 'run1'}

\item{path}{root directory of the run folder where .cov file resides}

\item{extension}{extension of the file to parse. Defaults to 'cov' but could also be 'cor'}

\item{clear_zip}{should the unzipped cov file be removed when done (defaults to T)}

\item{quiet}{suppresses the messages (defaults to not suppressing)}
}
\value{
List of covariance matrices as data frames
}
\description{
Unzips the zipped *.cov file and parses the elements for each estimation method into a data frame
}
\examples{

getOption("qpExampleDir")
run1.covmat =  nm_covmat_extract(run = "example1", path = getOption("qpExampleDir"))
names(run1.covmat)
names(run1.covmat[[1]])
lapply(run1.covmat, dim)

}
\seealso{
Other nm: 
\code{\link{extract_varcov}()},
\code{\link{nm_cov_table}()},
\code{\link{nm_unzip}()},
\code{\link{read_nm_qp}()},
\code{\link{sample_uncert_qp}()}
}
\concept{nm}
