% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_lst_by_est_method.r
\name{nm_lst_by_est_method.character}
\alias{nm_lst_by_est_method.character}
\title{Split NONMEM Output Listed by Estimation Method for Character}
\usage{
\method{nm_lst_by_est_method}{character}(
  x,
  directory = getOption("nmDir", getwd()),
  extension = c(getOption("nmFileLst", "ext"), getOption("nmFileLst", "lst")),
  nested = c(TRUE, FALSE),
  quiet = TRUE,
  clean = TRUE,
  clue = "\\\\$PROB",
  ...
)
}
\arguments{
\item{x}{character: a file path, or the name of a run in 'directory', or lst content}

\item{directory}{character: optional location of x}

\item{extension}{length-two character: extensions identifying ext and lst files}

\item{nested}{length-two logical: are the ext and lst files nested within the run}

\item{quiet}{logical: flag for displaying intermediate output}

\item{clean}{should the unzipped file be removed when done (defaults to TRUE)}

\item{clue}{a text fragment expected to occur in cov files}

\item{...}{passed arguments}
}
\value{
data.frame
}
\description{
Splits NONMEM output and put it in a list for each estimation method for Character
}
\examples{
library(magrittr)
library(dplyr)
dir <- getOption("qpExampleDir")
options(nmDir = dir)
nm_lst_by_est_method('run103')
nm_lst_by_est_method('run103') \%>\% lapply(., function(x) x[1:10])
}
\seealso{
Other lst: 
\code{\link{find_diag_template}()},
\code{\link{load_modeled_dataset}()},
\code{\link{nm_lst_by_est_method}()},
\code{\link{nm_lst_by_est_method.lst}()},
\code{\link{nm_methods}()},
\code{\link{nm_methods.character}()},
\code{\link{nm_methods.lst}()},
\code{\link{nm_methods_short}()},
\code{\link{nm_parse_control_stream}()},
\code{\link{nm_parse_dollar_data}()},
\code{\link{nm_shrinkage}()},
\code{\link{nm_version}()},
\code{\link{nm_version.character}()},
\code{\link{nm_version.lst}()},
\code{\link{read_lst}()},
\code{\link{read_lst.character}()},
\code{\link{read_lst.lst}()}
}
\concept{lst}
