% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_lst_by_est_method.r
\name{nm_lst_by_est_method.lst}
\alias{nm_lst_by_est_method.lst}
\title{Split NONMEM Output Listed by Estimation Method for LST}
\usage{
\method{nm_lst_by_est_method}{lst}(x, ...)
}
\arguments{
\item{x}{object of disptach of class 'lst'}

\item{...}{passed arguments}
}
\description{
Splits NONMEM output and put it in a list for each estimation method for LST
}
\seealso{
Other lst: 
\code{\link{find_diag_template}()},
\code{\link{load_modeled_dataset}()},
\code{\link{nm_lst_by_est_method}()},
\code{\link{nm_lst_by_est_method.character}()},
\code{\link{nm_methods}()},
\code{\link{nm_methods.character}()},
\code{\link{nm_methods.lst}()},
\code{\link{nm_methods_short}()},
\code{\link{nm_parse_control_stream}()},
\code{\link{nm_parse_dollar_data}()},
\code{\link{nm_shrinkage}()},
\code{\link{nm_version}()},
\code{\link{nm_version.character}()},
\code{\link{nm_version.lst}()},
\code{\link{read_lst}()},
\code{\link{read_lst.character}()},
\code{\link{read_lst.lst}()}
}
\concept{lst}
