% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_methods.r
\name{nm_methods.lst}
\alias{nm_methods.lst}
\title{Get NONMEM Estimation Methods for LST}
\usage{
\method{nm_methods}{lst}(x, ...)
}
\arguments{
\item{x}{lst (see \code{\link{read_lst}})}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Gets NONMEM estimation methods from the LST file.
}
\seealso{
Other lst: 
\code{\link{find_diag_template}()},
\code{\link{load_modeled_dataset}()},
\code{\link{nm_lst_by_est_method}()},
\code{\link{nm_lst_by_est_method.character}()},
\code{\link{nm_lst_by_est_method.lst}()},
\code{\link{nm_methods}()},
\code{\link{nm_methods.character}()},
\code{\link{nm_methods_short}()},
\code{\link{nm_parse_control_stream}()},
\code{\link{nm_parse_dollar_data}()},
\code{\link{nm_shrinkage}()},
\code{\link{nm_version}()},
\code{\link{nm_version.character}()},
\code{\link{nm_version.lst}()},
\code{\link{read_lst}()},
\code{\link{read_lst.character}()},
\code{\link{read_lst.lst}()}
}
\concept{lst}
