% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_params_table.r
\name{nm_params_table_short.character}
\alias{nm_params_table_short.character}
\title{Get Abbreviated Parameter Table for Character}
\usage{
\method{nm_params_table_short}{character}(
  x,
  selection = c("parameter", "estimate", "RSE", "CI95", "fix", "shrinksd", "symbol",
    "unit", "transform", "label"),
  ...
)
}
\arguments{
\item{x}{character, e.g. a run name like 'run01'}

\item{selection}{character: allowed columns}

\item{...}{passed to \code{\link{nm_params_table.character}} and \code{\link{nm_params_table_short.ext}}}
}
\value{
data.frame of class 'nmpartab'
}
\description{
Shortens the output of \code{\link{nm_params_table}} (based on EXT file) to essential information.
See also \code{\link{get_est_table}} which uses \pkg{pmxTools}
to construct a parameter table from the XML.
See also \code{\link[nonmemica]{partab}} from \pkg{nonmemica} which does similar.
}
\examples{
library(magrittr)
library(dplyr)
options(nmDir = getOption("qpExampleDir"))
"run103" \%>\% nm_params_table
"run103" \%>\% nm_params_table_short
"run103" \%>\% nm_params_table \%>\% nm_params_table_short
}
\seealso{
Other ext: 
\code{\link{as.list.ext}()},
\code{\link{as.matrix.ext}()},
\code{\link{find_diag}()},
\code{\link{format.nmpartab}()},
\code{\link{format.nmpartab_}()},
\code{\link{get_ext}()},
\code{\link{get_ext.character}()},
\code{\link{get_ext.ext}()},
\code{\link{get_ext.extList}()},
\code{\link{get_ofv.ext}()},
\code{\link{get_ofv.nonmem}()},
\code{\link{lol}()},
\code{\link{nm_condition_number}()},
\code{\link{nm_condition_number.default}()},
\code{\link{nm_condition_number.ext}()},
\code{\link{nm_params_table}()},
\code{\link{nm_params_table.character}()},
\code{\link{nm_params_table_short}()},
\code{\link{nm_params_table_short.ext}()},
\code{\link{nm_params_table_short.nmpartab}()},
\code{\link{read_ext}()},
\code{\link{read_ext.character}()},
\code{\link{read_ext.ext}()},
\code{\link{summary.extList}()}
}
\concept{ext}
