% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_parse_control_stream.r
\name{nm_parse_control_stream}
\alias{nm_parse_control_stream}
\title{Parse NONMEM control stream}
\usage{
nm_parse_control_stream(
  x,
  directory = getOption("nmDir", getwd()),
  extension = c(getOption("nmFileExt", "ext"), getOption("nmFileMod", "mod")),
  nested = c(TRUE, FALSE),
  quiet = TRUE,
  clean = TRUE,
  clue = "\\\\$PROB",
  ...
)
}
\arguments{
\item{x}{character: object of dispatch}

\item{directory}{character: optional location of x}

\item{extension}{length-two character: extensions identifying ext and lst files}

\item{nested}{length-two logical: are the ext and lst files nested within the run}

\item{quiet}{logical: flag for displaying intermediate output}

\item{clean}{should the unzipped file be removed when done (defaults to TRUE)}

\item{clue}{a text fragment expected to occur in cov files}

\item{...}{passed arguments}
}
\value{
list
}
\description{
Parses NONMEM control stream and list all elements into sections
}
\examples{
library(magrittr)

"run103" \%>\% nm_parse_control_stream
nm_parse_control_stream("example1", directory = getOption("qpExampleDir"),extension=c("ext","ctl"))

}
\seealso{
Other lst: 
\code{\link{coldex}()},
\code{\link{find_diag_template}()},
\code{\link{load_modeled_dataset}()},
\code{\link{nm_lst_by_est_method}()},
\code{\link{nm_lst_by_est_method.character}()},
\code{\link{nm_lst_by_est_method.lst}()},
\code{\link{nm_methods}()},
\code{\link{nm_methods.character}()},
\code{\link{nm_methods.lst}()},
\code{\link{nm_methods_short}()},
\code{\link{nm_parse_dollar_data}()},
\code{\link{nm_shrinkage}()},
\code{\link{nm_version}()},
\code{\link{nm_version.character}()},
\code{\link{nm_version.lst}()},
\code{\link{read_lst}()},
\code{\link{read_lst.character}()},
\code{\link{read_lst.lst}()},
\code{\link{rowdex}()}
}
\concept{lst}
