% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_parse_dollar_data.r
\name{nm_parse_dollar_data}
\alias{nm_parse_dollar_data}
\title{Parse the NONMEM List File Data Record}
\usage{
nm_parse_dollar_data(
  x,
  directory = getOption("nmDir", getwd()),
  extension = getOption("nmFileLst", "lst"),
  quiet = TRUE,
  clean = TRUE,
  nested = FALSE,
  clue = "\\\\$PROB",
  remove_at_sign = TRUE,
  ...
)
}
\arguments{
\item{x}{character: a file path, or the name of a run in 'directory', or lst content}

\item{directory}{character: optional location of x}

\item{extension}{character: extension identifying lst files}

\item{quiet}{logical: flag for displaying intermediate output}

\item{clean}{should the unzipped file be removed when done (defaults to TRUE)}

\item{nested}{is the lst file nested within the run directory?}

\item{clue}{a text fragment expected to occur in lst files}

\item{remove_at_sign}{whether to remove the 'at' symbol}

\item{...}{passed arguments}
}
\value{
list
}
\description{
Parses the NONMEM 'lst' file data record.
}
\examples{
  library(magrittr)
  library(nonmemica)
  options(nmDir = getOption('qpExampleDir'))
  'run100' \%>\% as.model \%$\% data \%>\% writeLines
  'run100' \%>\% datafile \%>\% suppressWarnings
# 'run100' \%>\% ignored # datafile not actually present
  'run100' \%>\% nm_parse_dollar_data

}
\seealso{
Other lst: 
\code{\link{coldex}()},
\code{\link{find_diag_template}()},
\code{\link{load_modeled_dataset}()},
\code{\link{nm_lst_by_est_method}()},
\code{\link{nm_lst_by_est_method.character}()},
\code{\link{nm_lst_by_est_method.lst}()},
\code{\link{nm_methods}()},
\code{\link{nm_methods.character}()},
\code{\link{nm_methods.lst}()},
\code{\link{nm_methods_short}()},
\code{\link{nm_parse_control_stream}()},
\code{\link{nm_shrinkage}()},
\code{\link{nm_version}()},
\code{\link{nm_version.character}()},
\code{\link{nm_version.lst}()},
\code{\link{read_lst}()},
\code{\link{read_lst.character}()},
\code{\link{read_lst.lst}()},
\code{\link{rowdex}()}
}
\concept{lst}
\keyword{internal}
