% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_process_scm.r
\name{nm_process_scm}
\alias{nm_process_scm}
\title{Parse SCM Output}
\usage{
nm_process_scm(directory, scmlog = "scmlog.txt", ...)
}
\arguments{
\item{directory}{full directory path where scmlog files reside}

\item{scmlog}{name of scm log of interest}

\item{...}{ignored}
}
\value{
data.frame
}
\description{
Reads and processes SCM output file such that the output can be directly copied into Microsoft products and LaTeX typesetting.
}
\examples{
library(magrittr)
library(dplyr)
library(knitr)
library(yamlet)
path = file.path(getOption("qpExampleDir"),"scm_example2")
nm_process_scm(path)
nm_process_scm(path) \%>\% 
  filter(chosen == 1) \%>\%                      # summary
  select(step, model, direction, pvalue) \%>\%
  alias \%>\%                                    # labels as column names
  resolve \%>\%                                  # substitute guide values 
  kable
 # model:
path \%>\% nm_process_scm \%>\% filter(chosen == 1) \%>\%
   select(model)
}
