% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_runrecord.r
\name{nm_runrecord}
\alias{nm_runrecord}
\title{Compile NONMEM Run Record}
\usage{
nm_runrecord(
  x,
  directory = getOption("nmDir", getwd()),
  extension = c(getOption("nmFileExt", "ext"), getOption("nmFileLst", "lst")),
  nested = c(TRUE, FALSE),
  index = NULL,
  ...
)
}
\arguments{
\item{x}{character vector of run names, e.g. c('run1','run2')}

\item{directory}{path to working directory}

\item{extension}{length-two character: extensions identifying ext and lst files}

\item{nested}{length-two logical: whether ext and lst files are nested in run directory}

\item{index}{integer indicating choice of estimation method (defaults to last)}

\item{...}{ignored}
}
\value{
data.frame
}
\description{
Compiles a NONMEM run record.
}
\examples{
library(magrittr)
options(nmDir = getOption('qpExampleDir'))
nm_runrecord(runc(100:105))
}
\seealso{
Other runrecord: 
\code{\link{get_xpose_tables}()},
\code{\link{nm_runrecord_graph}()}
}
\concept{runrecord}
