% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_shrinkage.r
\name{nm_shrinkage}
\alias{nm_shrinkage}
\title{Get NONMEM Shrinkage Estimates}
\usage{
nm_shrinkage(
  x,
  directory = getOption("nmDir", getwd()),
  extension = c(getOption("nmFileLst", "ext"), getOption("nmFileLst", "lst")),
  nested = c(TRUE, FALSE),
  quiet = TRUE,
  clean = TRUE,
  clue = "\\\\$PROB",
  keep_stat = "SD",
  keep_type = "ETA",
  ...
)
}
\arguments{
\item{x}{character: a file path, or the name of a run in 'directory', or lst content}

\item{directory}{character: optional location of x}

\item{extension}{length-two character: extensions identifying ext and lst files}

\item{nested}{length-two logical: are the ext and lst files nested within the run}

\item{quiet}{logical: flag for displaying intermediate output}

\item{clean}{should the unzipped file be removed when done (defaults to TRUE)}

\item{clue}{a text fragment expected to occur in cov files}

\item{keep_stat}{a text fragment indicating whether to take shrinkage variance ('VR') of standard deviation ('SD', default).}

\item{keep_type}{a text fragment indicating whether to take shrinkage 'EBV' or 'ETA' (default)}

\item{...}{passed arguments}
}
\value{
data.frame
}
\description{
Gets NONMEM shrinkage estimates from the LST file.
}
\examples{
dir <- getOption('qpExampleDir')
options(nmDir = dir)
# nm_shrinkage('example1') # old-style lst file not supported
nm_shrinkage('run103')
nm_shrinkage('run105')
}
\seealso{
Other lst: 
\code{\link{coldex}()},
\code{\link{find_diag_template}()},
\code{\link{load_modeled_dataset}()},
\code{\link{nm_lst_by_est_method}()},
\code{\link{nm_lst_by_est_method.character}()},
\code{\link{nm_lst_by_est_method.lst}()},
\code{\link{nm_methods}()},
\code{\link{nm_methods.character}()},
\code{\link{nm_methods.lst}()},
\code{\link{nm_methods_short}()},
\code{\link{nm_parse_control_stream}()},
\code{\link{nm_parse_dollar_data}()},
\code{\link{nm_version}()},
\code{\link{nm_version.character}()},
\code{\link{nm_version.lst}()},
\code{\link{read_lst}()},
\code{\link{read_lst.character}()},
\code{\link{read_lst.lst}()},
\code{\link{rowdex}()}
}
\concept{lst}
