% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_version.r
\name{nm_version.character}
\alias{nm_version.character}
\title{Get the NONMEM Version for Character}
\usage{
\method{nm_version}{character}(
  x,
  directory = getOption("nmDir", getwd()),
  extension = getOption("nmFileLst", "lst"),
  quiet = TRUE,
  clean = TRUE,
  nested = FALSE,
  clue = "\\\\$PROB",
  ...
)
}
\arguments{
\item{x}{character: a file path, or the name of a run in 'directory', or lst content}

\item{directory}{character: optional location of x}

\item{extension}{character: extension identifying lst files}

\item{quiet}{logical: flag for displaying intermediate output}

\item{clean}{should the unzipped file be removed when done (defaults to TRUE)}

\item{nested}{is the file nested within the run directory?}

\item{clue}{a text fragment expected to occur in lst files}

\item{...}{passed arguments}
}
\value{
length-one character of class 'lst'
}
\description{
Get the NONMEM Version for class character using \code{\link{read_lst}}.
}
\examples{
library(magrittr)
options(nmDir = getOption('qpExampleDir'))
'example1' \%>\% nm_version
'run100' \%>\% nm_version
}
\seealso{
Other lst: 
\code{\link{find_diag_template}()},
\code{\link{load_modeled_dataset}()},
\code{\link{nm_lst_by_est_method}()},
\code{\link{nm_lst_by_est_method.character}()},
\code{\link{nm_lst_by_est_method.lst}()},
\code{\link{nm_methods}()},
\code{\link{nm_methods.character}()},
\code{\link{nm_methods.lst}()},
\code{\link{nm_methods_short}()},
\code{\link{nm_parse_control_stream}()},
\code{\link{nm_parse_dollar_data}()},
\code{\link{nm_shrinkage}()},
\code{\link{nm_version}()},
\code{\link{nm_version.lst}()},
\code{\link{read_lst}()},
\code{\link{read_lst.character}()},
\code{\link{read_lst.lst}()}
}
\concept{lst}
