% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_version.r
\name{nm_version.lst}
\alias{nm_version.lst}
\title{Get the NONMEM Version for Lst}
\usage{
\method{nm_version}{lst}(x, ...)
}
\arguments{
\item{x}{character: of class lst}

\item{...}{ignored}
}
\description{
Get the NONMEM Version for class lst (see also \code{\link{read_lst}}).
}
\seealso{
Other lst: 
\code{\link{find_diag_template}()},
\code{\link{load_modeled_dataset}()},
\code{\link{nm_lst_by_est_method}()},
\code{\link{nm_lst_by_est_method.character}()},
\code{\link{nm_lst_by_est_method.lst}()},
\code{\link{nm_methods}()},
\code{\link{nm_methods.character}()},
\code{\link{nm_methods.lst}()},
\code{\link{nm_methods_short}()},
\code{\link{nm_parse_control_stream}()},
\code{\link{nm_parse_dollar_data}()},
\code{\link{nm_shrinkage}()},
\code{\link{nm_version}()},
\code{\link{nm_version.character}()},
\code{\link{read_lst}()},
\code{\link{read_lst.character}()},
\code{\link{read_lst.lst}()}
}
\concept{lst}
