% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odwg_classification.r
\name{odwg}
\alias{odwg}
\title{Calculate ODWG Classification}
\usage{
odwg(ast, bili, ULNAST = 40, ULNBILI = 17.104, ...)
}
\arguments{
\item{ast}{AST in same units as ULNAST}

\item{bili}{bilirubin in same units as ULNBILI}

\item{ULNAST}{upper limit of normal for AST in IU (default)}

\item{ULNBILI}{upper limit of normal for bilirubin in umol/L (default)}

\item{...}{ignored}
}
\value{
integer
}
\description{
Calculates ODWG classification from values of AST and bilirubin.
}
\examples{
library(magrittr)
library(dplyr)
x <- data.frame(id = 1:4, AST = c(39, 41, 39, 39), BILI = c(17, 17, 26, 52))
x \%>\% mutate(ODWG = odwg(AST, BILI))
}
\seealso{
Other basic: 
\code{\link{asNumeric}()},
\code{\link{compare_classes}()},
\code{\link{doubleStack}()},
\code{\link{doubleStack.decorated}()},
\code{\link{doubleStack.default}()},
\code{\link{extract_character}()},
\code{\link{extract_number}()},
\code{\link{fxs}()},
\code{\link{geocv}()},
\code{\link{geomean}()},
\code{\link{isMissing}()},
\code{\link{isNumeric}()},
\code{\link{logit}()},
\code{\link{logit_inv}()},
\code{\link{lunique}()},
\code{\link{make_symmetric}()},
\code{\link{odwg_classification.data.frame}()},
\code{\link{rgenpois}()},
\code{\link{rnorm_by_id}()},
\code{\link{runc}()},
\code{\link{sample_by_id}()},
\code{\link{sunique}()}
}
\concept{basic}
