% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odwg_classification.r
\name{odwg_classification.data.frame}
\alias{odwg_classification.data.frame}
\title{Calculate ODWG Classification for Data Frame}
\usage{
\method{odwg_classification}{data.frame}(x, ULNAST = 40, ULNBILI = 17.104, ...)
}
\arguments{
\item{x}{data.frame}

\item{ULNAST}{upper limit of normal for AST in IU (default)}

\item{ULNBILI}{upper limit of normal for bilirubin umol/L (default)}

\item{...}{ignored}
}
\value{
data.frame, with additional column ODWG (integer, 0-3)
}
\description{
Calculates ODWG classification for data frame.
}
\examples{
library(magrittr)
x <- data.frame(id = 1:4, AST = c(39, 41, 39, 39), BILI = c(17, 17, 26, 52))
x \%>\% odwg_classification
}
\seealso{
Other basic: 
\code{\link{asNumeric}()},
\code{\link{compare_classes}()},
\code{\link{doubleStack}()},
\code{\link{doubleStack.decorated}()},
\code{\link{doubleStack.default}()},
\code{\link{extract_character}()},
\code{\link{extract_number}()},
\code{\link{fxs}()},
\code{\link{geocv}()},
\code{\link{geomean}()},
\code{\link{isMissing}()},
\code{\link{isNumeric}()},
\code{\link{logit}()},
\code{\link{logit_inv}()},
\code{\link{lunique}()},
\code{\link{make_symmetric}()},
\code{\link{odwg}()},
\code{\link{rgenpois}()},
\code{\link{rnorm_by_id}()},
\code{\link{runc}()},
\code{\link{sample_by_id}()},
\code{\link{sunique}()}
}
\concept{basic}
