% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odwg_classification.r
\name{odwg_classification}
\alias{odwg_classification}
\title{Calculate ODWG Classification}
\usage{
odwg_classification(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Calculates ODWG classification.  Generic, with method \code{\link{odwg_classification.data.frame}}.
}
