% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_1comp_1abs.r
\name{pk_1comp_1abs}
\alias{pk_1comp_1abs}
\title{Check for missing values}
\usage{
pk_1comp_1abs(dose, tob, parms)
}
\arguments{
\item{dose}{value for dose level}

\item{tob}{numeric vector describing time of observation}

\item{parms}{numeric vector consisting of clearance, volume of distribution and absorption rate constant (in this order). The vector might be named but this is not required.}
}
\value{
A numeric vector of concentration predictions
}
\description{
Create a concentration prediction for a 1 compartment open model with first order absorption
}
\examples{
pk_1comp_1abs(dose=1000,tob=seq(0,48,1),parms=c(3,10,0.05))
}
\seealso{
Other modelpred: 
\code{\link{eff_1comp_1abs}()}
}
\concept{modelpred}
