% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tidyvpc_qp.r
\name{plot.qpvpc}
\alias{plot.qpvpc}
\title{Plot 'qpvpc'}
\usage{
\method{plot}{qpvpc}(
  x,
  point.col = 1,
  point.shape = 1,
  point.size = 0.75,
  point.alpha = 0.3,
  medPI.col = "#4CB54F",
  PI.alpha = 0.25,
  lsize = 0.5,
  scales = "free_y",
  add.rug = TRUE,
  rug.size = 0.5,
  rug.sides = "t",
  yaml = NULL,
  xLabel = NULL,
  yLabel = NULL,
  plot.bin.method = "xpose",
  facet_ncol = NULL,
  ...
)
}
\arguments{
\item{x}{'qpvpc' (inherits 'tidyvpcobj')}

\item{point.col}{passed to geom_point()}

\item{point.shape}{passed to geom_point()}

\item{point.size}{passed to geom_point()}

\item{point.alpha}{passed to geom_point()}

\item{medPI.col}{passed to scale_fill_manual()}

\item{PI.alpha}{passed to geom_ribbon()}

\item{lsize}{passed to geom_line() for quantiles}

\item{scales}{passed to facet_wrap()}

\item{add.rug}{passed to geom_rug()}

\item{rug.size}{passed to geom_rug()}

\item{rug.sides}{passed to geom_rug()}

\item{yaml}{passed to \code{\link[yamlet]{redecorate}} as 'meta'}

\item{xLabel}{passed to xlab()}

\item{yLabel}{passed to ylab()}

\item{plot.bin.method}{if "xpose", suppresses PI.alpha}

\item{facet_ncol}{passed to facet_wrap()}

\item{...}{ignored}
}
\value{
ggplot
}
\description{
Plots a tidy vpc using qP defaults.
}
\examples{
\dontrun{
library(magrittr)
library(ggplot2)
x \%>\% as_qpvpc \%>\% plot
}

}
\seealso{
Other tidyvpc: 
\code{\link{as_qpvpc}()},
\code{\link{as_qpvpc.tidyvpcobj}()}
}
\concept{tidyvpc}
