% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_colors.r
\name{qp_colors}
\alias{qp_colors}
\title{Identify qp_colors}
\usage{
qp_colors(n = 22L)
}
\arguments{
\item{n}{length-one integer for number of colors to return}
}
\value{
character
}
\description{
Identifies qp_colors.
}
\examples{
qp_colors
}
