% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_colors.r
\name{qp_colors_sorted}
\alias{qp_colors_sorted}
\title{Identify sorted qp_colors}
\usage{
qp_colors_sorted(n = 22L)
}
\arguments{
\item{n}{length-one integer for number of colors to return}
}
\value{
character
}
\description{
Identifies sorted qp_colors.
}
\examples{
qp_colors
}
