% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bootstrap.r
\name{read_bootstrap}
\alias{read_bootstrap}
\title{Reads NONMEM bootstrap results for processing}
\usage{
read_bootstrap(
  directory = getOption("nmDir"),
  filename = "raw_results1.csv",
  structure.filename = "raw_results_structure",
  parlist = c("theta", "omega", "sigma", "ofv"),
  quiet = F
)
}
\arguments{
\item{directory}{Directory where bootstrap summary files reside}

\item{filename}{default Filename of the bootstrap results summary file created by PsN bootstrap}

\item{structure.filename}{Filename of structure file created by PsN bootstrap}

\item{parlist}{character string detailing what components of the bootstrap you want}

\item{quiet}{suppresses a message on how many bootstrap repliactes were read}
}
\value{
A stuctured and named list with data frames with containing bootstrap results
}
\description{
Reads raw files files of a completed NONMEM bootstrap
}
\examples{
\dontrun{
  library(magrittr)
  boot_dir = getOption("nmDir") \%>\% file.path(., "bootstrap")
  fn_boot = "raw_results_run105bs.csv"
  bootstrap = read_bootstrap(boot_dir, filename = fn_boot)
  bootstrap$boot_base 
  bootstrap$boot_base \%>\% do.call("cbind",.) \%>\% hist()
  bootstrap$structure
  bootstrap$directory
  
  bootstrap \%>\% tabulate_bootstrap()
}

}
