% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_coi.r
\name{read_coi.character}
\alias{read_coi.character}
\title{Read NONMEM COI Results for Character}
\usage{
\method{read_coi}{character}(
  x,
  directory = getOption("nmDir", getwd()),
  extension = getOption("nmFilecoi", "coi"),
  quiet = TRUE,
  clean = TRUE,
  nested = TRUE,
  clue = "TABLE NO.",
  ...
)
}
\arguments{
\item{x}{character: a file path, or the name of a run in 'directory', or coi content}

\item{directory}{character: optional location of x}

\item{extension}{character: extension identifying coi files}

\item{quiet}{logical: flag for displaying intermediate output}

\item{clean}{should the unzipped file be removed when done (defaults to TRUE)}

\item{nested}{is the file nested within the run directory?}

\item{clue}{a text fragment expected to occur in coi files}

\item{...}{passed arguments}
}
\value{
length-one character of class 'coi'
}
\description{
Reads NONMEM COI results for class character using \code{\link{read_gen.character}}.
}
\examples{
library(magrittr)
options(nmDir = getOption('qpExampleDir'))
'example1' \%>\% read_coi
'example1.coi' \%>\% read_coi
path <- file.path(getOption('qpExampleDir'), 'example1', 'example1.coi.7z')
path
file.exists(path)
path \%>\% read_coi
path \%>\% read_coi \%>\% as.matrix
path \%>\% get_coimat # same
'example1' \%>\% get_coimat
'example1' \%>\% get_covmat
'example1' \%>\% get_cormat
}
\seealso{
Other coi: 
\code{\link{get_coimat}()},
\code{\link{get_coimat.character}()},
\code{\link{get_coimat.coi}()},
\code{\link{read_coi}()},
\code{\link{read_coi.coi}()}
}
\concept{coi}
