% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_coi.r
\name{read_coi.coi}
\alias{read_coi.coi}
\title{Read NONMEM COI results for 'coi'}
\usage{
\method{read_coi}{coi}(x, ...)
}
\arguments{
\item{x}{class 'coi'}

\item{...}{ignored}
}
\value{
character vector of class 'coi'
}
\description{
Reads NONMEM COI results for class 'coi'. 
Just returns its argument, which is already the result of reading COI.
}
\seealso{
Other coi: 
\code{\link{get_coimat}()},
\code{\link{get_coimat.character}()},
\code{\link{get_coimat.coi}()},
\code{\link{read_coi}()},
\code{\link{read_coi.character}()}
}
\concept{coi}
\keyword{internal}
