% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_coi.r
\name{read_coi}
\alias{read_coi}
\title{Read NONMEM COI Results}
\usage{
read_coi(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed}
}
\description{
Reads NONMEM COI results.
Generic, with method \code{\link{read_coi.character}}
}
\seealso{
Other coi: 
\code{\link{get_coimat}()},
\code{\link{get_coimat.character}()},
\code{\link{get_coimat.coi}()},
\code{\link{read_coi.character}()},
\code{\link{read_coi.coi}()}
}
\concept{coi}
\keyword{internal}
