% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cor.r
\name{read_cor.character}
\alias{read_cor.character}
\title{Read NONMEM COR Results for Character}
\usage{
\method{read_cor}{character}(
  x,
  directory = getOption("nmDir", getwd()),
  extension = getOption("nmFilecor", "cor"),
  quiet = TRUE,
  clean = TRUE,
  nested = TRUE,
  clue = "TABLE NO.",
  ...
)
}
\arguments{
\item{x}{character: a file path, or the name of a run in 'directory', or cor content}

\item{directory}{character: optional location of x}

\item{extension}{character: extension identifying cor files}

\item{quiet}{logical: flag for displaying intermediate output}

\item{clean}{should the unzipped file be removed when done (defaults to TRUE)}

\item{nested}{is the file nested within the run directory?}

\item{clue}{a text fragment expected to occur in cor files}

\item{...}{passed arguments}
}
\value{
length-one character of class 'cor'
}
\description{
Reads NONMEM COR results for class character using \code{\link{read_gen.character}}.
}
\examples{
library(magrittr)
options(nmDir = getOption('qpExampleDir'))
'example1' \%>\% read_cor
'example1.cor' \%>\% read_cor
path <- file.path(getOption('qpExampleDir'), 'example1', 'example1.cor.7z')
path
file.exists(path)
path \%>\% read_cor
path \%>\% read_cor \%>\% as.matrix
path \%>\% get_cormat # same
'example1' \%>\% get_cormat
'example1' \%>\% get_covmat
'example1' \%>\% get_coimat
}
\seealso{
Other cor: 
\code{\link{get_cormat}()},
\code{\link{get_cormat.character}()},
\code{\link{get_cormat.cor}()},
\code{\link{read_cor}()},
\code{\link{read_cor.cor}()}
}
\concept{cor}
