% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cor.r
\name{read_cor.cor}
\alias{read_cor.cor}
\title{Read NONMEM COR results for 'cor'}
\usage{
\method{read_cor}{cor}(x, ...)
}
\arguments{
\item{x}{class 'cor'}

\item{...}{ignored}
}
\value{
character vector of class 'cor'
}
\description{
Reads NONMEM COR results for class 'cor'. 
Just returns its argument, which is already the result of reading COR.
}
\seealso{
Other cor: 
\code{\link{get_cormat}()},
\code{\link{get_cormat.character}()},
\code{\link{get_cormat.cor}()},
\code{\link{read_cor}()},
\code{\link{read_cor.character}()}
}
\concept{cor}
\keyword{internal}
