% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cor.r
\name{read_cor}
\alias{read_cor}
\title{Read NONMEM COR Results}
\usage{
read_cor(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed}
}
\description{
Reads NONMEM COR results.
Generic, with method \code{\link{read_cor.character}}
}
\seealso{
Other cor: 
\code{\link{get_cormat}()},
\code{\link{get_cormat.character}()},
\code{\link{get_cormat.cor}()},
\code{\link{read_cor.character}()},
\code{\link{read_cor.cor}()}
}
\concept{cor}
\keyword{internal}
