% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cov.r
\name{read_cov.character}
\alias{read_cov.character}
\title{Read NONMEM COV Results for Character}
\usage{
\method{read_cov}{character}(
  x,
  directory = getOption("nmDir", getwd()),
  extension = getOption("nmFileCov", "cov"),
  quiet = TRUE,
  clean = TRUE,
  nested = TRUE,
  clue = "TABLE NO.",
  ...
)
}
\arguments{
\item{x}{character: a file path, or the name of a run in 'directory', or cov content}

\item{directory}{character: optional location of x}

\item{extension}{character: extension identifying cov files}

\item{quiet}{logical: flag for displaying intermediate output}

\item{clean}{should the unzipped file be removed when done (defaults to TRUE)}

\item{nested}{is the file nested within the run directory?}

\item{clue}{a text fragment expected to occur in cov files}

\item{...}{passed arguments}
}
\value{
length-one character of class 'cov'
}
\description{
Reads NONMEM COV results for class character using \code{\link{read_gen.character}}.
}
\examples{
library(magrittr)
options(nmDir = getOption('qpExampleDir'))
'example1' \%>\% read_cov
'example1.cov' \%>\% read_cov
path <- file.path(getOption('qpExampleDir'), 'example1', 'example1.cov.7z')
path
file.exists(path)
path \%>\% read_cov
path \%>\% read_cov \%>\% as.matrix
path \%>\% get_covmat # same
'example1' \%>\% get_covmat
'example1' \%>\% get_cormat
'example1' \%>\% get_coimat
}
\seealso{
Other cov: 
\code{\link{get_covmat}()},
\code{\link{get_covmat.character}()},
\code{\link{get_covmat.cov}()},
\code{\link{read_cov}()},
\code{\link{read_cov.cov}()}
}
\concept{cov}
