% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cov.r
\name{read_cov.cov}
\alias{read_cov.cov}
\title{Read NONMEM COV results for 'cov'}
\usage{
\method{read_cov}{cov}(x, ...)
}
\arguments{
\item{x}{class 'cov'}

\item{...}{ignored}
}
\value{
character vector of class 'cov'
}
\description{
Reads NONMEM COV results for class 'cov'. 
Just returns its argument, which is already the result of reading COV.
}
\seealso{
Other cov: 
\code{\link{get_covmat}()},
\code{\link{get_covmat.character}()},
\code{\link{get_covmat.cov}()},
\code{\link{read_cov}()},
\code{\link{read_cov.character}()}
}
\concept{cov}
\keyword{internal}
