% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cov.r
\name{read_cov}
\alias{read_cov}
\title{Read NONMEM COV Results}
\usage{
read_cov(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed}
}
\description{
Reads NONMEM COV results.
Generic, with method \code{\link{read_cov.character}}
}
\seealso{
Other cov: 
\code{\link{get_covmat}()},
\code{\link{get_covmat.character}()},
\code{\link{get_covmat.cov}()},
\code{\link{read_cov.character}()},
\code{\link{read_cov.cov}()}
}
\concept{cov}
\keyword{internal}
