% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ext.r
\name{read_ext.ext}
\alias{read_ext.ext}
\title{Read NONMEM EXT results}
\usage{
\method{read_ext}{ext}(x, ...)
}
\arguments{
\item{x}{class 'ext'}

\item{...}{ignored}
}
\value{
character of class 'ext'
}
\description{
Reads NONMEM EXT results for class 'ext'. 
Just returns its argument, which is already the result of reading EXT.
}
\seealso{
Other ext: 
\code{\link{as.list.ext}()},
\code{\link{as.matrix.ext}()},
\code{\link{find_diag}()},
\code{\link{format.nmpartab}()},
\code{\link{format.nmpartab_}()},
\code{\link{get_ext}()},
\code{\link{get_ext.character}()},
\code{\link{get_ext.ext}()},
\code{\link{get_ext.extList}()},
\code{\link{get_ofv.ext}()},
\code{\link{get_ofv.nonmem}()},
\code{\link{lol}()},
\code{\link{nm_condition_number}()},
\code{\link{nm_condition_number.default}()},
\code{\link{nm_condition_number.ext}()},
\code{\link{nm_params_table}()},
\code{\link{nm_params_table.character}()},
\code{\link{nm_params_table_short}()},
\code{\link{nm_params_table_short.character}()},
\code{\link{nm_params_table_short.ext}()},
\code{\link{nm_params_table_short.nmpartab}()},
\code{\link{read_ext}()},
\code{\link{read_ext.character}()},
\code{\link{summary.extList}()}
}
\concept{ext}
\keyword{internal}
