% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gen.r
\name{read_gen.character}
\alias{read_gen.character}
\title{Read NONMEM File for Character}
\usage{
\method{read_gen}{character}(
  x,
  directory = getOption("nmDir", getwd()),
  extension = NULL,
  quiet = TRUE,
  clean = TRUE,
  nested = TRUE,
  clue = NULL,
  ...
)
}
\arguments{
\item{x}{character: a file path, or the name of a run in 'directory', or file content}

\item{directory}{character: optional location of x}

\item{extension}{character: a file extension indicating desired file type}

\item{quiet}{logical: flag for displaying intermediate output}

\item{clean}{should the occur in a temporary directory?}

\item{nested}{is the file nested within the run directory?}

\item{clue}{character: a text fragment that ought to occur in files with this extension}

\item{...}{passed arguments}
}
\value{
length-one character of file contents, classified like 'extension'
}
\description{
Reads NONMEM file for class character.
If length of x is greater than one, it is collapsed.
If x is a file, possibly in \code{directory}, it is read with \code{\link{read_gen.file}}.
If x implies a zipped file, possibly in \code{directory}, it is unzipped first.
If x is a directory, possibly in \code{directory}, it is read with \code{\link{read_gen.run}}.
If an unzipped version of the file already exists in the target directory (local or temporary) it is removed before unzipping proceeds.
}
\examples{
 library(magrittr)
 options(nmDir = getOption('qpExampleDir'))
 'example1' \%>\% read_gen('xml')
 'example1' \%>\% read_gen('cov')
 'example1' \%>\% read_gen('lst') # also unnested!
 'example1.lst' \%>\% read_gen(nested = FALSE) # also nested!
#'example1' \%>\% read_gen('ctl') # should not work!
 'example1' \%>\% read_gen('ctl', nested = FALSE)
 'example1' \%>\% read_gen('ext')
 'example1.cov' \%>\% read_gen('ext') # mistake!
 'example1.cov' \%>\% read_gen('cov')
 'example1.cov' \%>\% read_gen
 'example1.cov' \%>\% file.path(getOption('qpExampleDir'), 'example1', .) \%>\% read_gen
 'example1.cov.7z' \%>\% read_gen

}
\seealso{
Other gen: 
\code{\link{read_gen}()},
\code{\link{read_gen.file}()},
\code{\link{read_gen.run}()},
\code{\link{read_gen.text}()}
}
\concept{gen}
