% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gen.r
\name{read_gen.file}
\alias{read_gen.file}
\title{Read NONMEM File as File}
\usage{
\method{read_gen}{file}(x, extension, directory, quiet, clean, nested, ...)
}
\arguments{
\item{x}{character: a file path}

\item{extension}{character: a file extension indicating desired file type}

\item{directory}{character: optional location of x}

\item{quiet}{logical: flag for displaying intermediate output}

\item{clean}{should the unzipped file be removed when done (defaults to TRUE)}

\item{nested}{is the file nested within the run directory?}

\item{...}{passed arguments}
}
\value{
length-one character of file contents, classified like 'extension'
}
\description{
Reads NONMEM file for class 'file'
}
\seealso{
Other gen: 
\code{\link{read_gen}()},
\code{\link{read_gen.character}()},
\code{\link{read_gen.run}()},
\code{\link{read_gen.text}()}
}
\concept{gen}
\keyword{internal}
