% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gen.r
\name{read_gen}
\alias{read_gen}
\title{Read Generic NONMEM File}
\usage{
read_gen(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed}
}
\description{
Reads generic NONMEM file with particular extension.
Generic, with method \code{\link{read_gen.character}}
}
\seealso{
Other gen: 
\code{\link{read_gen.character}()},
\code{\link{read_gen.file}()},
\code{\link{read_gen.run}()},
\code{\link{read_gen.text}()}
}
\concept{gen}
\keyword{internal}
