% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gen.r
\name{read_gen.text}
\alias{read_gen.text}
\title{Read NONMEM File as Text}
\usage{
\method{read_gen}{text}(x, extension, ...)
}
\arguments{
\item{x}{length-one character with text content}

\item{extension}{character: a file extension indicating desired file type}

\item{...}{ignored}
}
\value{
character
}
\description{
Reads NONMEM file for class 'text': length-one character with text content.
}
\seealso{
Other gen: 
\code{\link{read_gen}()},
\code{\link{read_gen.character}()},
\code{\link{read_gen.file}()},
\code{\link{read_gen.run}()}
}
\concept{gen}
\keyword{internal}
