% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lst.r
\name{read_lst.character}
\alias{read_lst.character}
\title{Read NONMEM LST Results for Character}
\usage{
\method{read_lst}{character}(
  x,
  directory = getOption("nmDir", getwd()),
  extension = getOption("nmFileLst", "lst"),
  quiet = TRUE,
  clean = TRUE,
  nested = FALSE,
  clue = "\\\\$PROB",
  ...
)
}
\arguments{
\item{x}{character: a file path, or the name of a run in 'directory', or lst content}

\item{directory}{character: optional location of x}

\item{extension}{character: extension identifying lst files}

\item{quiet}{logical: flag for displaying intermediate output}

\item{clean}{should the unzipped file be removed when done (defaults to TRUE)}

\item{nested}{is the file nested within the run directory?}

\item{clue}{a text fragment expected to occur in lst files}

\item{...}{passed arguments}
}
\value{
length-one character of class 'lst'
}
\description{
Reads NONMEM LST results for class character using \code{\link{read_gen.character}}.
}
\examples{
library(magrittr)
options(nmDir = getOption('qpExampleDir'))
'example1' \%>\% read_lst
'example1.lst' \%>\% read_lst
'example1.lst' \%>\% read_lst(nested = TRUE) # different file! (if it exists)
path <- file.path(getOption('qpExampleDir'), 'example1', 'example1.lst.7z')
path
stopifnot(file.exists(path))
path \%>\% read_lst
'run100' \%>\% nm_shrinkage
}
\seealso{
Other lst: 
\code{\link{coldex}()},
\code{\link{find_diag_template}()},
\code{\link{load_modeled_dataset}()},
\code{\link{nm_lst_by_est_method}()},
\code{\link{nm_lst_by_est_method.character}()},
\code{\link{nm_lst_by_est_method.lst}()},
\code{\link{nm_methods}()},
\code{\link{nm_methods.character}()},
\code{\link{nm_methods.lst}()},
\code{\link{nm_methods_short}()},
\code{\link{nm_parse_control_stream}()},
\code{\link{nm_parse_dollar_data}()},
\code{\link{nm_shrinkage}()},
\code{\link{nm_version}()},
\code{\link{nm_version.character}()},
\code{\link{nm_version.lst}()},
\code{\link{read_lst}()},
\code{\link{read_lst.lst}()},
\code{\link{rowdex}()}
}
\concept{lst}
