% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lst.r
\name{read_lst.lst}
\alias{read_lst.lst}
\title{Read NONMEM LST results}
\usage{
\method{read_lst}{lst}(x, ...)
}
\arguments{
\item{x}{class 'lst'}

\item{...}{ignored}
}
\value{
character vector of class 'lst'
}
\description{
Reads NONMEM LST results for class 'lst'. 
Just returns its argument, which is already the result of reading LST.
}
\seealso{
Other lst: 
\code{\link{find_diag_template}()},
\code{\link{load_modeled_dataset}()},
\code{\link{nm_lst_by_est_method}()},
\code{\link{nm_lst_by_est_method.character}()},
\code{\link{nm_lst_by_est_method.lst}()},
\code{\link{nm_methods}()},
\code{\link{nm_methods.character}()},
\code{\link{nm_methods.lst}()},
\code{\link{nm_methods_short}()},
\code{\link{nm_parse_control_stream}()},
\code{\link{nm_parse_dollar_data}()},
\code{\link{nm_shrinkage}()},
\code{\link{nm_version}()},
\code{\link{nm_version.character}()},
\code{\link{nm_version.lst}()},
\code{\link{read_lst}()},
\code{\link{read_lst.character}()}
}
\concept{lst}
\keyword{internal}
