% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lst.r
\name{read_lst}
\alias{read_lst}
\title{Read NONMEM LST Results}
\usage{
read_lst(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed}
}
\description{
Reads NONMEM LST results.
Generic, with method \code{\link{read_lst.character}}
}
\seealso{
Other lst: 
\code{\link{coldex}()},
\code{\link{find_diag_template}()},
\code{\link{load_modeled_dataset}()},
\code{\link{nm_lst_by_est_method}()},
\code{\link{nm_lst_by_est_method.character}()},
\code{\link{nm_lst_by_est_method.lst}()},
\code{\link{nm_methods}()},
\code{\link{nm_methods.character}()},
\code{\link{nm_methods.lst}()},
\code{\link{nm_methods_short}()},
\code{\link{nm_parse_control_stream}()},
\code{\link{nm_parse_dollar_data}()},
\code{\link{nm_shrinkage}()},
\code{\link{nm_version}()},
\code{\link{nm_version.character}()},
\code{\link{nm_version.lst}()},
\code{\link{read_lst.character}()},
\code{\link{read_lst.lst}()},
\code{\link{rowdex}()}
}
\concept{lst}
\keyword{internal}
