% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mod.r
\name{read_mod.mod}
\alias{read_mod.mod}
\title{Read NONMEM MOD results}
\usage{
\method{read_mod}{mod}(x, ...)
}
\arguments{
\item{x}{class 'mod'}

\item{...}{ignored}
}
\value{
character vector of class 'mod'
}
\description{
Reads NONMEM MOD results for class 'mod'. 
Just returns its argument, which is already the result of reading MOD.
}
\seealso{
Other mod: 
\code{\link{read_mod}()},
\code{\link{read_mod.character}()}
}
\concept{mod}
\keyword{internal}
