% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mod.r
\name{read_mod}
\alias{read_mod}
\title{Read NONMEM MOD Results}
\usage{
read_mod(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed}
}
\description{
Reads NONMEM MOD results.
Generic, with method \code{\link{read_mod.character}}
}
\seealso{
Other mod: 
\code{\link{read_mod.character}()},
\code{\link{read_mod.mod}()}
}
\concept{mod}
\keyword{internal}
