% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm.r
\name{read_nm.character}
\alias{read_nm.character}
\title{Read NONMEM Results for Character}
\usage{
\method{read_nm}{character}(
  x,
  directory = getOption("nmDir", getwd()),
  quiet = FALSE,
  clean = TRUE,
  ...
)
}
\arguments{
\item{x}{character: a file path, or the name of a run in 'directory', or xml content}

\item{directory}{character: optional location of x}

\item{quiet}{logical: flag for displaying intermediate output}

\item{clean}{should the unzipped file be removed when done (defaults to TRUE)}

\item{...}{passed arguments}
}
\value{
list of class 'nonmem'
}
\description{
Reads NONMEM results for class character.
If length of x is greater than one, it is collapsed.
If x is a file, possibly in \code{directory}, it is read with \code{\link{read_nm.file}}.
If x is a directory, possibly in \code{directory}, it is read with \code{\link{read_nm.run}}.
Otherwise an attempt is made to read it with \code{\link{read_nm.xml}}.
See also \code{\link[pmxTools]{read_nm}} on which this depends.
}
\examples{
nm <- read_nm("example1", directory = getOption("qpExampleDir"))
names(nm)
names(nm$nonmem)
names(nm$nonmem$problem)
names(nm$nonmem$problem$estimation)

# consider multiple ways of addressing an xml file

dir <- getOption("qpExampleDir")
run <- 'example1'
zip <- 'example1.xml.7z'
path <- file.path(dir, run, file)
file.exists(path)
to <- nm_unzip(zip.filename = path)
xml <- readLines(path)
}
\seealso{
Other nm: 
\code{\link{as_pmx_canonical}()},
\code{\link{definitions.nonmem}()},
\code{\link{extract_varcov}()},
\code{\link{get_est_table.default}()},
\code{\link{get_est_table.nonmem}()},
\code{\link{get_est_table}()},
\code{\link{nm_cov_table}()},
\code{\link{nm_covmat_extract}()},
\code{\link{nm_unzip}()},
\code{\link{read_nm.file}()},
\code{\link{read_nm.run}()},
\code{\link{read_nm.xml}()},
\code{\link{read_nm}()},
\code{\link{sample_uncert_qp}()}
}
\concept{nm}
