% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm.r
\name{read_nm.file}
\alias{read_nm.file}
\title{Read NONMEM Results as File}
\usage{
\method{read_nm}{file}(x, directory = getwd(), quiet = FALSE, ...)
}
\arguments{
\item{x}{length-one character: path to an xml file, possibly in \code{directory}}

\item{directory}{character: optional location of x}

\item{quiet}{logical: flag for displaying intermediate output}

\item{...}{passed arguments}
}
\value{
list of class 'nonmem'
}
\description{
Reads NONMEM results for class 'file': a path to an xml file.
See also \code{\link[pmxTools]{read_nm}}.
}
\seealso{
Other nm: 
\code{\link{as_pmx_canonical}()},
\code{\link{definitions.nonmem}()},
\code{\link{extract_varcov}()},
\code{\link{get_est_table.default}()},
\code{\link{get_est_table.nonmem}()},
\code{\link{get_est_table}()},
\code{\link{nm_cov_table}()},
\code{\link{nm_covmat_extract}()},
\code{\link{nm_unzip}()},
\code{\link{read_nm.character}()},
\code{\link{read_nm.run}()},
\code{\link{read_nm.xml}()},
\code{\link{read_nm}()},
\code{\link{sample_uncert_qp}()}
}
\concept{nm}
