% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm.r
\name{read_nm}
\alias{read_nm}
\title{Read NONMEM Results}
\usage{
read_nm(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed}
}
\description{
Reads NONMEM results.
Generic, with method \code{\link{read_nm.character}}
}
\seealso{
Other nm: 
\code{\link{as_pmx_canonical}()},
\code{\link{definitions.nonmem}()},
\code{\link{extract_varcov}()},
\code{\link{get_est_table.default}()},
\code{\link{get_est_table.nonmem}()},
\code{\link{get_est_table}()},
\code{\link{nm_cov_table}()},
\code{\link{nm_covmat_extract}()},
\code{\link{nm_unzip}()},
\code{\link{read_nm.character}()},
\code{\link{read_nm.file}()},
\code{\link{read_nm.run}()},
\code{\link{read_nm.xml}()},
\code{\link{sample_uncert}()}
}
\concept{nm}
