% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm.r
\name{read_nm.run}
\alias{read_nm.run}
\title{Read the entire XML output of a Converged NONMEM Run}
\usage{
\method{read_nm}{run}(
  x,
  directory = getOption("nmDir", getwd()),
  quiet = FALSE,
  clean = TRUE,
  ...
)
}
\arguments{
\item{x}{character string in the form of 'run1'}

\item{directory}{parent directory of the run folder where .xml file resides}

\item{quiet}{suppresses the messages (defaults to not suppressing)}

\item{clean}{should the unzipped file be removed when done (defaults to TRUE, but ignored if file was already present)}

\item{...}{passed arguments}
}
\value{
nonmem: List of all elements of the XML output of a NONMEM model run
}
\description{
Extracts the output of a completed NONMEM run that is stored in an XML file,
possibly 7-zipped.
}
\examples{
nm <- read_nm("example1", directory = getOption("qpExampleDir"))
names(nm)
names(nm$nonmem)
names(nm$nonmem$problem)
names(nm$nonmem$problem$estimation)
}
\seealso{
Other nm: 
\code{\link{as_pmx_canonical}()},
\code{\link{definitions.nonmem}()},
\code{\link{extract_varcov}()},
\code{\link{get_est_table.default}()},
\code{\link{get_est_table.nonmem}()},
\code{\link{get_est_table}()},
\code{\link{nm_cov_table}()},
\code{\link{nm_covmat_extract}()},
\code{\link{nm_unzip}()},
\code{\link{read_nm.character}()},
\code{\link{read_nm.file}()},
\code{\link{read_nm.xml}()},
\code{\link{read_nm}()},
\code{\link{sample_uncert_qp}()}
}
\concept{nm}
