% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm.r
\name{read_nm.xml}
\alias{read_nm.xml}
\title{Read NONMEM Results as XML}
\usage{
\method{read_nm}{xml}(x, quiet = FALSE, ...)
}
\arguments{
\item{x}{length-one character with xml content}

\item{quiet}{logical: flag for displaying intermediate output}

\item{...}{passed arguments}
}
\value{
list of class 'nonmem'
}
\description{
Reads NONMEM results for class 'xml': length-one character with xml content.
See also \code{\link[pmxTools]{read_nm}}.
}
\seealso{
Other nm: 
\code{\link{as_pmx_canonical}()},
\code{\link{definitions.nonmem}()},
\code{\link{extract_varcov}()},
\code{\link{get_est_table.default}()},
\code{\link{get_est_table.nonmem}()},
\code{\link{get_est_table}()},
\code{\link{nm_cov_table}()},
\code{\link{nm_covmat_extract}()},
\code{\link{nm_unzip}()},
\code{\link{read_nm.character}()},
\code{\link{read_nm.file}()},
\code{\link{read_nm.run}()},
\code{\link{read_nm}()},
\code{\link{sample_uncert_qp}()}
}
\concept{nm}
