% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm_qp.r
\name{read_nm_qp.character}
\alias{read_nm_qp.character}
\title{Read NONMEM Results for Character}
\usage{
\method{read_nm_qp}{character}(
  x,
  directory = getOption("nmDir", getwd()),
  extension = getOption("nmFileXml", "xml"),
  quiet = TRUE,
  clean = TRUE,
  nested = TRUE,
  clue = "</nm:output>",
  ...
)
}
\arguments{
\item{x}{character: a file path, or the name of a run in 'directory', or xml content}

\item{directory}{character: optional location of x}

\item{extension}{character: extension identifying cov files}

\item{quiet}{logical: flag for displaying intermediate output}

\item{clean}{should the unzipped file be removed when done (defaults to TRUE)}

\item{nested}{is the file nested within the run directory?}

\item{clue}{a text fragment expected to occur in cov files}

\item{...}{passed arguments}
}
\value{
list of class 'nonmem'
}
\description{
Reads NONMEM results for class character using \code{\link{read_gen.character}}.
See also \code{\link[pmxTools]{read_nm}}.
}
\examples{
nm <- read_nm_qp("example1", directory = getOption("qpExampleDir"))
names(nm)
names(nm$nonmem)
names(nm$nonmem$problem)
names(nm$nonmem$problem$estimation)

# consider multiple ways of addressing an xml file

dir <- getOption("qpExampleDir")
options(nmDir = dir)
run <- 'example1'
file <- 'example1.xml'
zipped <- 'example1.xml.7z'
path <- file.path(dir, run, zipped)
file.exists(path)

nm1 <- read_nm_qp(run)
nm2 <- read_nm_qp(run, directory = dir)
nm3 <- read_nm_qp(file)
nm4 <- read_nm_qp(file, directory = dir)
nm5 <- read_nm_qp(path)
}
\seealso{
Other nm: 
\code{\link{definitions.nonmem}()},
\code{\link{extract_varcov}()},
\code{\link{get_est_table}()},
\code{\link{get_est_table.default}()},
\code{\link{get_est_table.nonmem}()},
\code{\link{get_ofv}()},
\code{\link{get_ofv.character}()},
\code{\link{nm_cov_table}()},
\code{\link{nm_unzip}()},
\code{\link{read_nm_qp}()},
\code{\link{read_nm_qp.nonmem}()},
\code{\link{sample_uncert_qp}()}
}
\concept{nm}
