% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm_qp.r
\name{read_nm_qp.nonmem}
\alias{read_nm_qp.nonmem}
\title{Read NONMEM results}
\usage{
\method{read_nm_qp}{nonmem}(x, ...)
}
\arguments{
\item{x}{class 'nonmem'}

\item{...}{ignored}
}
\value{
character vector of class 'nonmem'
}
\description{
Reads NONMEM results for class 'nonmem'. 
Just returns its argument, which is already the result of reading nonmem.
}
\seealso{
Other nm: 
\code{\link{definitions.nonmem}()},
\code{\link{extract_varcov}()},
\code{\link{get_est_table}()},
\code{\link{get_est_table.default}()},
\code{\link{get_est_table.nonmem}()},
\code{\link{get_ofv}()},
\code{\link{get_ofv.character}()},
\code{\link{nm_cov_table}()},
\code{\link{nm_unzip}()},
\code{\link{read_nm_qp}()},
\code{\link{read_nm_qp.character}()},
\code{\link{sample_uncert_qp}()}
}
\concept{nm}
\keyword{internal}
