% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm_qp.r
\name{read_nm_qp}
\alias{read_nm_qp}
\title{Read the entire XML output of a Converged NONMEM Run}
\usage{
read_nm_qp(
  run,
  path = getOption("nmDir", getwd()),
  clear_zip = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{run}{character string in the form of 'run1'}

\item{path}{root directory of the run folder where .cov file resides}

\item{clear_zip}{should the unzipped cov file be removed when done (defaults to T)}

\item{quiet}{suppresses the messages (defaults to not suppressing)}
}
\value{
List of all elements of the XML output of a NONMEM model run
}
\description{
Extracts the output of a completed NONMEM run that is stored in an XML file.
}
\examples{

xml1 = read_nm_qp(run="example1", path = getOption("qpExampleDir"))
names(xml1)
names(xml1$nonmem)
names(xml1$nonmem$problem)
names(xml1$nonmem$problem$estimation)
}
\seealso{
Other nm: 
\code{\link{extract_varcov}()},
\code{\link{nm_cov_table}()},
\code{\link{nm_covmat_extract}()},
\code{\link{nm_unzip}()},
\code{\link{sample_uncert_qp}()}
}
\concept{nm}
