% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnorm_by_id.r
\name{rnorm_by_id}
\alias{rnorm_by_id}
\title{Group-wise sampling from normal distribution}
\usage{
rnorm_by_id(id, mean = 0, sd = 1)
}
\arguments{
\item{id}{vector with group information}

\item{mean}{mode of the normal distribution}

\item{sd}{standard deviation of the normal distribution}
}
\value{
samples of a normal distribution, one for each level of \code{group}
}
\description{
Populates 1 sample from normal distribution for each group (id or subject, or trial, or other)
}
\examples{
df = data.frame(id=rep(1:3,ea=4))
sam = rnorm_by_id(df$id, mean = 0, sd = 1)
tapply(sam,df$id, unique)
}
\seealso{
Other basic: 
\code{\link{asNumeric}()},
\code{\link{compare_classes}()},
\code{\link{doubleStack}()},
\code{\link{doubleStack.decorated}()},
\code{\link{doubleStack.default}()},
\code{\link{extract_character}()},
\code{\link{extract_number}()},
\code{\link{fxs}()},
\code{\link{geocv}()},
\code{\link{geomean}()},
\code{\link{isMissing}()},
\code{\link{isNumeric}()},
\code{\link{logit}()},
\code{\link{logit_inv}()},
\code{\link{lunique}()},
\code{\link{make_symmetric}()},
\code{\link{odwg}()},
\code{\link{odwg_classification.data.frame}()},
\code{\link{rgenpois}()},
\code{\link{runc}()},
\code{\link{sample_by_id}()},
\code{\link{sunique}()}
}
\concept{basic}
