% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_diag_template.r
\name{rowdex}
\alias{rowdex}
\title{Row indices for a half matrix}
\usage{
rowdex(x, ...)
}
\arguments{
\item{x}{integer or numeric giving the order (one dimension) of a square matrix}

\item{...}{ignored}
}
\description{
Return the row indices for the lower triangle (including diagonal)
of a square matrix, given its order (length of diagonal).
}
\examples{
rowdex(0)
rowdex(3)

}
\seealso{
Other lst: 
\code{\link{coldex}()},
\code{\link{find_diag_template}()},
\code{\link{load_modeled_dataset}()},
\code{\link{nm_lst_by_est_method}()},
\code{\link{nm_lst_by_est_method.character}()},
\code{\link{nm_lst_by_est_method.lst}()},
\code{\link{nm_methods}()},
\code{\link{nm_methods.character}()},
\code{\link{nm_methods.lst}()},
\code{\link{nm_methods_short}()},
\code{\link{nm_parse_control_stream}()},
\code{\link{nm_parse_dollar_data}()},
\code{\link{nm_shrinkage}()},
\code{\link{nm_version}()},
\code{\link{nm_version.character}()},
\code{\link{nm_version.lst}()},
\code{\link{read_lst}()},
\code{\link{read_lst.character}()},
\code{\link{read_lst.lst}()}
}
\concept{lst}
\keyword{internal}
