% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runc.r
\name{runc}
\alias{runc}
\title{Glue run rootname and number together}
\usage{
runc(x, run = "run")
}
\arguments{
\item{x}{numeric vector}

\item{run}{prefix of run number, defaults to "run"}
}
\value{
data.frame
}
\description{
Pastes run rootname and number together.
}
\examples{
library(magrittr)
1:10 \%>\% runc
c(878,1021,1030:1035) \%>\% runc

}
\seealso{
Other basic: 
\code{\link{asNumeric}()},
\code{\link{compare_classes}()},
\code{\link{doubleStack}()},
\code{\link{doubleStack.decorated}()},
\code{\link{doubleStack.default}()},
\code{\link{extract_character}()},
\code{\link{extract_number}()},
\code{\link{fxs}()},
\code{\link{geocv}()},
\code{\link{geomean}()},
\code{\link{isMissing}()},
\code{\link{isNumeric}()},
\code{\link{logit}()},
\code{\link{logit_inv}()},
\code{\link{lunique}()},
\code{\link{make_symmetric}()},
\code{\link{odwg}()},
\code{\link{odwg_classification.data.frame}()},
\code{\link{rgenpois}()},
\code{\link{rnorm_by_id}()},
\code{\link{sample_by_id}()},
\code{\link{sunique}()}
}
\concept{basic}
