% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_uncert_qp.r
\name{sample_uncert_qp}
\alias{sample_uncert_qp}
\title{Draw Samples from the Quasi-Posterior Parameter Estimate Distribution.}
\usage{
sample_uncert_qp(
  run,
  path = getOption("nmDir"),
  n = 10,
  seed = 1234,
  est_step = NULL,
  clear_zip = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{run}{character string in the form of 'run1'}

\item{path}{root directory of the run folder where .cov file resides}

\item{n}{sample size of the requested distribution}

\item{seed}{random seed for simulation}

\item{est_step}{estimation number - defaults to the last one}

\item{clear_zip}{remove the xml and cov file after unzipping? Defaults to TRUE}

\item{quiet}{suppresses the messages (defaults to not suppressing)}
}
\value{
List of covariance matrices
}
\description{
Read point estimates and the variance-covariance matrix of a model run and create a distribution of parameter estimates forming a sample of the quasi-posterior parameter distribution
}
\examples{

unc1 = sample_uncert_qp(run = "example1", path = getOption("qpExampleDir"))
unc1 = sample_uncert_qp(run = "example1", path = getOption("qpExampleDir"), quiet=TRUE, n=100)
dim(unc1)
}
\seealso{
Other nm: 
\code{\link{as_pmx_canonical}()},
\code{\link{definitions.nonmem}()},
\code{\link{extract_varcov}()},
\code{\link{get_est_table.default}()},
\code{\link{get_est_table.nonmem}()},
\code{\link{get_est_table}()},
\code{\link{nm_cov_table}()},
\code{\link{nm_covmat_extract}()},
\code{\link{nm_unzip}()},
\code{\link{read_nm.character}()},
\code{\link{read_nm.file}()},
\code{\link{read_nm.run}()},
\code{\link{read_nm.xml}()},
\code{\link{read_nm}()}
}
\concept{nm}
