% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap.r
\name{snap}
\alias{snap}
\title{Coerce Values to Nearest of Candidates}
\usage{
snap(x, rule = 1, left = TRUE, ...)
}
\arguments{
\item{x}{numeric}

\item{rule}{a vector of (finite numeric) candidates, or a single value giving candidate interval on the real number line}

\item{left}{whether to return the lesser of two equidistant candidates}

\item{...}{ignored}
}
\value{
numeric
}
\description{
For each value in a numeric vector, return the closest match from a vector of candidate values.
}
\details{
If rule is scalar, it must be positive; a rule will be constructed as a sequence of rule-spaced values that includes zero and includes values at least as extreme as the extremes of x. In some sense, this function is the complement to cut: whereas in cut one specifies the "breaks", with snap one specifies a set of "attractors" (breaks are the implied midpoints); both functions map their primary argument to a (possibly) smaller set of values.
Note: If length of x is zero, it is returned unmodified. If length of rule is zero, value is all NA.
}
\examples{
snap(c(0.0, 1.0, 1.2, 2.0, 2.9, 3))
snap(-10:10, 0.3)
times <- c(0,.25,.5,1,2,4,5,7,9,12,24)
snap(c(3,NA,5), c(2,4,6))
snap(c(3,NA,5), numeric(0))
}
