% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ext.r
\name{summary.extList}
\alias{summary.extList}
\title{Summarize 'extList'}
\usage{
\method{summary}{extList}(object, final = FALSE, ...)
}
\arguments{
\item{object}{class 'extList', as from \code{\link{as.list.ext}}}

\item{final}{logical: return just the iterations with numbers greater than zero}

\item{...}{ignored}
}
\value{
data.frame
}
\description{
Summarizes 'extList'. See \code{\link{as.list.ext}}.
}
\details{
The stochastic iterations of the SAEM analysis are given negative values. 
These are followed by positive iterations of the accumulation phase.
  * Iteration -1000000000 (negative one billion) indicates that this line 
  contains the final result (thetas, omegas, and sigmas, and objective function) 
  of the particular analysis. For BAYES analysis, this is the mean of the 
  non-negative iterations (stationary samples) listed before it.
  * Iteration -1000000001 indicates that this line contains the standard 
  errors of the final population parameters. For BAYES, it is the sample 
  standard deviation of the stationary samples.
  * Iteration -1000000002 indicates that this line contains the eigenvalues 
  of the correlation matrix of the variances of the final parameters.
  * Iteration -1000000003 indicates that this line contains the condition 
  number , lowest, highest, Eigen values of the correlation matrix of the 
  variances of the final parameters.
  * Iteration -1000000004 indicates this line contains the OMEGA and SIGMA 
  elements in standard deviation/correlation format.
  * Iteration -1000000005 indicates this line contains the standard errors to 
  the OMEGA and SIGMA elements in standard deviation/correlation format.
  * Iteration -1000000006 indicates 1 if parameter was fixed in estimation, 0 otherwise.
  * Iteration -1000000007 lists termination status (first item) followed by 
  termination codes. See I.55 $EST: Additional Output Files Produced under 
  root.xml (NM72) for interpreting the codes.
  * Iteration -1000000008 lists the partial derivative of the log likelihood 
  (-1/2 OFV) with respect to each estimated parameter. This may be useful for 
  using tests like the Lagrange multiplier test.
}
\seealso{
Other ext: 
\code{\link{as.list.ext}()},
\code{\link{as.matrix.ext}()},
\code{\link{find_diag}()},
\code{\link{format.nmpartab}()},
\code{\link{format.nmpartab_}()},
\code{\link{get_ext}()},
\code{\link{get_ext.character}()},
\code{\link{get_ext.ext}()},
\code{\link{get_ext.extList}()},
\code{\link{get_ofv.ext}()},
\code{\link{get_ofv.nonmem}()},
\code{\link{lol}()},
\code{\link{nm_condition_number}()},
\code{\link{nm_condition_number.default}()},
\code{\link{nm_condition_number.ext}()},
\code{\link{nm_params_table}()},
\code{\link{nm_params_table.character}()},
\code{\link{nm_params_table_short}()},
\code{\link{nm_params_table_short.character}()},
\code{\link{nm_params_table_short.ext}()},
\code{\link{nm_params_table_short.nmpartab}()},
\code{\link{read_ext}()},
\code{\link{read_ext.character}()},
\code{\link{read_ext.ext}()}
}
\concept{ext}
