% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_bootstrap.r
\name{tabulate_bootstrap}
\alias{tabulate_bootstrap}
\title{Create Parameter Table from NONMEM Bootstrap}
\usage{
tabulate_bootstrap(
  bootstrap,
  probs = 0.95,
  stat = list(central = median),
  digits = 3,
  latex = FALSE,
  align.dot = FALSE,
  transformations = NULL
)
}
\arguments{
\item{bootstrap}{A bootstrap object of class 'bootstrap' created by read_bootstrap.}

\item{probs}{The confidence interval coverage.}

\item{stat}{a list with at least option 'central' which is for a central tendency function of the distribution (e.g. median, mean)}

\item{digits}{number of significant digits in output; passed to \code{\link{formatted_signif}}}

\item{latex}{passed to \code{\link{formatted_signif}}}

\item{align.dot}{passed to \code{\link{formatted_signif}}}

\item{transformations}{a list with elements log or logit, that, in turn, are numeric vectors representating the THETA numbers that are logged or logit-transformed, respectively.}
}
\value{
A data.frame with the bootstrap estimate. Check the output as assumptions are used for assignment of column 'parameter' and these may not fully match for OMEGA and SIGMA, which is especially untested for complex BLOCK OMEGA structures, and when IOV is coded via REPLACE.
}
\description{
Creates parameter table from NONMEM bootstrap procedure.
}
\examples{
library(magrittr)
library(dplyr)
boot_dir = getOption("nmDir") \%>\% file.path(., "bootstrap")
fn_boot = "raw_results_run105bs.csv"
myBoot = read_bootstrap(
 directory = boot_dir,
 filename = fn_boot,
)
myBoot \%>\% tabulate_bootstrap
myBoot \%>\% tabulate_bootstrap(transformations=list(log=c(1,4),logit=c(6,7)))

# merge bootstrap results in
param105 = "run105" \%>\% nm_params_table() \%>\% lol
theBoots = myBoot \%>\% tabulate_bootstrap()
param105 \%<>\% left_join(theBoots \%>\% filter(parameter !="ofv") \%>\% select(parameter,b_central,b_CI))
param105 \%>\% select(parameter,estimate, se_est, CI95, shrinksd, b_central,b_CI)
}
