% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_latex.r
\name{to_latex.data.frame}
\alias{to_latex.data.frame}
\title{Convert Dataframe to Latex}
\usage{
\method{to_latex}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{ignored}
}
\description{
Converts 'data.frame' to latex.
Iterates across columns, calling to_latex().
}
\examples{
to_latex(Theoph)

}
\seealso{
Other latex: 
\code{\link{to_latex}()},
\code{\link{to_latex.default}()},
\code{\link{to_latex.nmpartab}()},
\code{\link{to_latex.nms_nonmem}()}
}
\concept{latex}
