% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_latex.r
\name{to_latex.default}
\alias{to_latex.default}
\title{Convert Anything to Latex}
\usage{
\method{to_latex}{default}(x, ...)
}
\arguments{
\item{x}{anything}

\item{...}{ignored}
}
\description{
Converts anything to latex.
By default, does nothing!
This lets us avoid unintentional coversions.
}
\seealso{
Other latex: 
\code{\link{to_latex}()},
\code{\link{to_latex.data.frame}()},
\code{\link{to_latex.nmpartab}()},
\code{\link{to_latex.nms_nonmem}()}
}
\concept{latex}
