% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_latex.r
\name{to_latex.nmpartab}
\alias{to_latex.nmpartab}
\title{Convert Parameter Tables to Latex}
\usage{
\method{to_latex}{nmpartab}(x, ...)
}
\arguments{
\item{x}{nmpartab}

\item{...}{ignored arguments}
}
\description{
Converts 'nmpartab' (list of parameter tables) to latex.
calls to_latex() for each member.
}
\seealso{
Other latex: 
\code{\link{to_latex}()},
\code{\link{to_latex.data.frame}()},
\code{\link{to_latex.default}()},
\code{\link{to_latex.nms_nonmem}()}
}
\concept{latex}
