% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_latex.r
\name{to_latex.nms_nonmem}
\alias{to_latex.nms_nonmem}
\title{Convert Nonmem Names to Latex}
\usage{
\method{to_latex}{nms_nonmem}(x, ...)
}
\arguments{
\item{x}{nms_nonmem}

\item{...}{ignored arguments}
}
\description{
Converts nms_nonmem to latex. 
Expects THETA(1), OMEGA(2,2), SIGMA(1,1) etc.
}
\seealso{
Other latex: 
\code{\link{to_latex}()},
\code{\link{to_latex.data.frame}()},
\code{\link{to_latex.default}()},
\code{\link{to_latex.nmpartab}()}
}
\concept{latex}
