% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_latex.r
\name{to_latex}
\alias{to_latex}
\title{Convert to Latex}
\usage{
to_latex(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Converts something to latex. Generic, with method \code{\link{to_latex.nms_nonmem}}.
}
\examples{
library(magrittr)
library(dplyr)
library(knitr)
library(yamlet)
options(nmDir = getOption("qpExampleDir"))
"run103" \%>\% 
  nm_params_table \%>\% 
  lol \%>\%
  format \%>\%
  to_latex
}
\seealso{
Other latex: 
\code{\link{to_latex.data.frame}()},
\code{\link{to_latex.default}()},
\code{\link{to_latex.nmpartab}()},
\code{\link{to_latex.nms_nonmem}()}
}
\concept{latex}
