% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.r
\name{trendpair.data.frame}
\alias{trendpair.data.frame}
\title{Plot Two Trendplots for Data.frame}
\usage{
\method{trendpair}{data.frame}(
  data,
  x1,
  y1,
  x2 = NULL,
  y2 = NULL,
  ...,
  alpha = 0.5,
  transform_x = "identity",
  transform_y = "identity",
  ref = 0,
  aspect = 1,
  ncol = 2
)
}
\arguments{
\item{data}{data.frame}

\item{x1}{bareword for column mapped to x axis 1}

\item{y1}{bareword for column mapped to y axis 1}

\item{x2}{bareword for column mapped to x axis 2 (defaults to x1 internally)}

\item{y2}{bareword for column mapped to y axis 2 (defaults to y1 internally)}

\item{...}{ignored}

\item{alpha}{passed to \code{\link{isoplot}}, can be length 2}

\item{transform_x}{passed to \code{\link{trendplot}}, can be length 2}

\item{transform_y}{passed to \code{\link{trendplot}}, can be length 2}

\item{ref}{passed to \code{\link{trendplot}}, can be length 2}

\item{aspect}{passed to \code{\link{trendplot}}, can be length 2}

\item{ncol}{number of columns (1 or 2):  2 gives side-by-side layout}
}
\value{
ggplot
}
\description{
Plots two trendplots for 'data.frame', typically side-by-side.
}
\examples{
library(magrittr)
library(yamlet)
x <- 'extdata/modeling/xanomeline-mod.csv.gz' \%>\% 
system.file(package = 'yamlet') \%>\% io_csv \%>\% filter(cp > 1) \%>\% resolve
x \%>\% trendpair(TIME, IWRES, TIME, IWRES, transform_x = c('identity', 'log10'))
}
\seealso{
Other plots: 
\code{\link{isopair}()},
\code{\link{isopair.data.frame}()},
\code{\link{isoplot}()},
\code{\link{isoplot.data.frame}()},
\code{\link{trendpair}()},
\code{\link{trendplot}()},
\code{\link{trendplot.data.frame}()}
}
\concept{plots}
\keyword{visualization}
