% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.r
\name{trendpair}
\alias{trendpair}
\title{Plot Two Trendplots}
\usage{
trendpair(data, ...)
}
\arguments{
\item{data}{object of dispatch}

\item{...}{passed}
}
\description{
Plots two trendplots.
Generic, with method \code{\link{trendpair.data.frame}}.
}
\seealso{
Other plots: 
\code{\link{isopair}()},
\code{\link{isopair.data.frame}()},
\code{\link{isoplot}()},
\code{\link{isoplot.data.frame}()},
\code{\link{trendpair.data.frame}()},
\code{\link{trendplot}()},
\code{\link{trendplot.data.frame}()}
}
\concept{plots}
\keyword{internal}
