% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.r
\name{trendplot.data.frame}
\alias{trendplot.data.frame}
\title{Plot a Trend for Dataframe}
\usage{
\method{trendplot}{data.frame}(
  data,
  x,
  y,
  ...,
  alpha = 0.5,
  transform_x = "identity",
  transform_y = "identity",
  ref = 0,
  aspect = 1,
  symmetric = transform_y == "identity"
)
}
\arguments{
\item{data}{data.frame}

\item{x}{bareword for column mapped to x axis}

\item{y}{bareword for column mapped to y axis}

\item{...}{ignored}

\item{alpha}{passed to \code{\link[ggplot2]{geom_point}}}

\item{transform_x}{passed to \code{\link[ggplot2]{scale_x_continuous}}}

\item{transform_y}{passed to \code{\link[ggplot2]{scale_x_continuous}}}

\item{ref}{passed to \code{\link[ggplot2]{geom_hline}} as \code{yintercept}}

\item{aspect}{passed to \code{\link[ggplot2]{theme}} as \code{aspect.ratio}}

\item{symmetric}{whether to enforce y-axis symmetry around zero, default if \code{transform = 'identity'}}
}
\value{
ggplot
}
\description{
Plots a trend for 'data.frame'.
By default adds y-axis symmetry.
Adds reference and trend lines.
Exposes alpha and transformation.
}
\examples{
library(magrittr)
library(yamlet)
x <- 'extdata/modeling/xanomeline-mod.csv.gz' \%>\% 
system.file(package = 'yamlet') \%>\% io_csv \%>\% filter(cp > 1) \%>\% resolve
x \%>\% trendplot(TIME, IWRES, transform_x = 'log10')

}
\seealso{
Other plots: 
\code{\link{isopair}()},
\code{\link{isopair.data.frame}()},
\code{\link{isoplot}()},
\code{\link{isoplot.data.frame}()},
\code{\link{trendpair}()},
\code{\link{trendpair.data.frame}()},
\code{\link{trendplot}()}
}
\concept{plots}
\keyword{visualization}
