% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.r
\name{trendplot}
\alias{trendplot}
\title{Plot a Trend}
\usage{
trendplot(data, ...)
}
\arguments{
\item{data}{object of dispatch}

\item{...}{passed}
}
\description{
Plots a trend.
Generic, with method \code{\link{trendplot.data.frame}}.
}
\seealso{
Other plots: 
\code{\link{isopair}()},
\code{\link{isopair.data.frame}()},
\code{\link{isoplot}()},
\code{\link{isoplot.data.frame}()},
\code{\link{trendpair}()},
\code{\link{trendpair.data.frame}()},
\code{\link{trendplot.data.frame}()}
}
\concept{plots}
\keyword{internal}
