% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm_qp.r
\name{as_pmx_canonical}
\alias{as_pmx_canonical}
\title{Convert Parameter Names to pmxTools Style}
\usage{
as_pmx_canonical(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{ignored}
}
\value{
character
}
\description{
Converts parameter names from nonmemica \code{link[nonmemica]{nms_canonical}}
to pmxTools style.  I.e., theta_01 becomes THETA1, and 
omega_01_01 becomes OM1,1.
}
\examples{
as_pmx_canonical(
 c(
   'theta_15',
   'omega_01_01',
   'omega_01_02',
   'sigma_01_01',
   'sigma_02_01'
 )
)

}
\seealso{
Other nm: 
\code{\link{definitions.nonmem}()},
\code{\link{extract_varcov}()},
\code{\link{get_est_table.default}()},
\code{\link{get_est_table.nonmem}()},
\code{\link{get_est_table}()},
\code{\link{nm_cov_table}()},
\code{\link{nm_covmat_extract}()},
\code{\link{nm_unzip}()},
\code{\link{read_nm_qp}()},
\code{\link{sample_uncert_qp}()}
}
\concept{nm}
