% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_number.r
\name{extract_number}
\alias{extract_number}
\title{Takes character string and parses the numeric input from it, if available.}
\usage{
extract_number(x)
}
\arguments{
\item{x}{A character or numeric string.}
}
\value{
A numeric vector if the input has multiple character blocks.
}
\description{
Takes character string and parses the numeric input from it, if available.
}
\section{Notes}{

If multiple blocks of numeric input exist wthin the input, only the first is parsed
}

\examples{
extract_number("c76")
extract_number("c76c7")
extract_number("BLOQ = 0.144")
extract_number("6.023+23")
stopifnot(extract_number("6.1973+01") == 61.973)
extract_number(c('7e1', '7e+1', '7E-1'))
extract_number(c('-7e1', '+7e+1', '-7E-1'))
extract_number('E7')
}
\seealso{
Other basic: 
\code{\link{asNumeric}()},
\code{\link{compare_classes}()},
\code{\link{doubleStack}()},
\code{\link{doubleStack.decorated}()},
\code{\link{doubleStack.default}()},
\code{\link{extract_character}()},
\code{\link{fxs}()},
\code{\link{geocv}()},
\code{\link{geomean}()},
\code{\link{isMissing}()},
\code{\link{isNumeric}()},
\code{\link{logit}()},
\code{\link{logit_inv}()},
\code{\link{lunique}()},
\code{\link{make_symmetric}()},
\code{\link{odwg}()},
\code{\link{odwg_classification.data.frame}()},
\code{\link{rgenpois}()},
\code{\link{rnorm_by_id}()},
\code{\link{runc}()},
\code{\link{sample_by_id}()},
\code{\link{sunique}()}
}
\concept{basic}
