% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxs.r
\name{fxs}
\alias{fxs}
\title{Fixed character length of a String}
\usage{
fxs(x, pattern = "0000")
}
\arguments{
\item{x}{string to format}

\item{pattern}{format to adhere to}
}
\value{
x reformatted according to pattern
}
\description{
Format string of fixed character length
}
\examples{
fxs(c(0,1,34, 50000), pattern = "    ")
fxs(c(0,1,34, 50000), "000000")
}
\seealso{
Other basic: 
\code{\link{asNumeric}()},
\code{\link{compare_classes}()},
\code{\link{doubleStack}()},
\code{\link{doubleStack.decorated}()},
\code{\link{doubleStack.default}()},
\code{\link{extract_character}()},
\code{\link{extract_number}()},
\code{\link{geocv}()},
\code{\link{geomean}()},
\code{\link{isMissing}()},
\code{\link{isNumeric}()},
\code{\link{logit}()},
\code{\link{logit_inv}()},
\code{\link{lunique}()},
\code{\link{make_symmetric}()},
\code{\link{odwg}()},
\code{\link{odwg_classification.data.frame}()},
\code{\link{rgenpois}()},
\code{\link{rnorm_by_id}()},
\code{\link{runc}()},
\code{\link{sample_by_id}()},
\code{\link{sunique}()}
}
\concept{basic}
