% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm_qp.r
\name{get_est_table.nonmem}
\alias{get_est_table.nonmem}
\title{Create a Table of Model Parameters for NONMEM}
\usage{
\method{get_est_table}{nonmem}(
  x,
  thetaLabels = NULL,
  omegaLabels = NULL,
  sigmaLabels = NULL,
  sigdig = 3,
  fields = getOption("fields", c("symbol", "unit", "transform", "label")),
  ...,
  meta = TRUE
)
}
\arguments{
\item{x}{object of dispatch}

\item{thetaLabels}{passed to \code{\link[pmxTools]{get_est_table}}}

\item{omegaLabels}{passed to \code{\link[pmxTools]{get_est_table}}}

\item{sigmaLabels}{passed to \code{\link[pmxTools]{get_est_table}}}

\item{sigdig}{passed to \code{\link[pmxTools]{get_est_table}}}

\item{fields}{character: names of semicolon-delimited fields to expect in control streams}

\item{...}{passed to \code{\link[pmxTools]{get_est_table}} and \code{\link[nonmemica]{definitions}}}

\item{meta}{whether to merge metadata (labels etc.) from control stream}
}
\description{
Creates a Table of Model Parameters for class 'nonmem'.
See also \code{\link{read_nm_qp}} which returns this class.
if \code{meta} is TRUE, an attempt is made
to supply metadata using markup in the nonmem control stream
(as saved in the xml) and the result of \code{definitions(x)}.
The value of \code{getOption('fields')} is highly relevant.
See also \code{\link[nonmemica]{psn_options}}.
}
\examples{
library(magrittr)
options(nmDir = getOption('qpExampleDir'))
'example1' \%>\% 
  read_nm_qp \%>\% 
  get_est_table \%$\% 
  Parameter

}
\seealso{
Other nm: 
\code{\link{definitions.nonmem}()},
\code{\link{extract_varcov}()},
\code{\link{get_est_table}()},
\code{\link{get_est_table.default}()},
\code{\link{get_ofv}()},
\code{\link{get_ofv.character}()},
\code{\link{nm_cov_table}()},
\code{\link{nm_unzip}()},
\code{\link{read_nm_qp}()},
\code{\link{read_nm_qp.character}()},
\code{\link{read_nm_qp.nonmem}()},
\code{\link{sample_uncert_qp}()}
}
\concept{nm}
