% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lunique.r
\name{lunique}
\alias{lunique}
\title{Number of unique values}
\usage{
lunique(x)
}
\arguments{
\item{x}{values to be uniqued}
}
\value{
Number of unique values in vector \code{x}
}
\description{
Derive the length (number) of unique values in a vector
}
\examples{
lunique(1:10)
lunique(c(1,1,2,2,3,3,4,4))
}
\seealso{
Other basic: 
\code{\link{asNumeric}()},
\code{\link{compare_classes}()},
\code{\link{doubleStack}()},
\code{\link{doubleStack.decorated}()},
\code{\link{doubleStack.default}()},
\code{\link{extract_character}()},
\code{\link{extract_number}()},
\code{\link{fxs}()},
\code{\link{geocv}()},
\code{\link{geomean}()},
\code{\link{isMissing}()},
\code{\link{isNumeric}()},
\code{\link{logit}()},
\code{\link{logit_inv}()},
\code{\link{make_symmetric}()},
\code{\link{odwg}()},
\code{\link{odwg_classification.data.frame}()},
\code{\link{rgenpois}()},
\code{\link{rnorm_by_id}()},
\code{\link{runc}()},
\code{\link{sample_by_id}()},
\code{\link{sunique}()}
}
\concept{basic}
