% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_cov_table.r
\name{nm_cov_table}
\alias{nm_cov_table}
\title{Double Stacking or Wide to Tall Transformation of Parameter and ETA Estimates with Covariates}
\usage{
nm_cov_table(
  x,
  ID = "ID",
  parameters = character(0),
  covariates = character(0),
  carryAlong = character(0),
  wide2tall = "none",
  n_cat = 10
)
}
\arguments{
\item{x}{NONMEM output tables as a data.frame from get.xpose.tables, or read_nmtables, or alike}

\item{ID}{the subject identifier in the NONMEM output}

\item{parameters}{Model parameter name vector}

\item{covariates}{covariates to integrate}

\item{carryAlong}{columns to carry along i the output}

\item{wide2tall}{either 'parameter' or 'eta'. Plural will be accepted too}

\item{n_cat}{number of unique values to qualify as a categorical covariate}
}
\value{
data frame with parValue, etaValue, covValue, and cov_class (continuous or categorical)
}
\description{
Double stack parameter estimates and ETA values with covariates into a single data frame
}
\examples{

library(magrittr)
library(tidyr)
library(dplyr)
options(nmDir=getOption('qpExampleDir'))
out = 'run103' \%>\% get_xpose_tables()
out \%<>\% nm_cov_table(parameters=c('CL', 'V', 'KA'),
                      covariates=c('WT'),
                      carryAlong=NULL
) 

out \%>\% nm_cov_table(parameters=c('CL', 'V'),
                     covariates=c('WT'),
                     carryAlong=NULL,
                     wide2tall='etas'
) 

out \%>\% nm_cov_table(parameters=c('CL', 'V', 'KA'),
                     covariates=c( 'WT'),
                     carryAlong=NULL,
                     wide2tall='parameter'
) 

library(ggplot2)
out = "run100" \%>\% get_xpose_tables 
out \%<>\% nm_cov_table(
  parameters=c("CL","V","KA"), 
  covariates = c("WT","SEX","ETN"), 
  wide2tall = "parameters"
)
outcon = out \%>\% 
  filter(cov_class=="continuous") \%>\% 
  mutate(covValue=as.numeric(covValue))
outcat = out \%>\% filter(cov_class=="categorical")
outcon \%>\%
  ggplot(aes(x=covValue,y=parValue)) + 
  geom_point() + 
  stat_smooth() + 
  facet_wrap(covariate~parameter, scales="free_y")
  
outcat \%>\%
  ggplot(aes(y=covValue,x=parValue)) + 
  geom_point() + 
  geom_boxplot(horizontal=TRUE) + 
  facet_grid(covariate~parameter, scales="free_x")
  
}
\seealso{
Other nm: 
\code{\link{definitions.nonmem}()},
\code{\link{extract_varcov}()},
\code{\link{get_est_table}()},
\code{\link{get_est_table.default}()},
\code{\link{get_est_table.nonmem}()},
\code{\link{get_ofv}()},
\code{\link{get_ofv.character}()},
\code{\link{nm_unzip}()},
\code{\link{read_nm_qp}()},
\code{\link{read_nm_qp.character}()},
\code{\link{read_nm_qp.nonmem}()},
\code{\link{sample_uncert_qp}()}
}
\concept{nm}
