% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_unzip.r
\name{nm_unzip}
\alias{nm_unzip}
\title{Unzip a NONMEM File}
\usage{
nm_unzip(
  run,
  rundir = normalizePath(getOption("nmDir", getwd())),
  path = file.path(rundir, run),
  extension = ".xml",
  filename = paste0(run, fxs(extension, ".   ")),
  zipped = file.path(path, filename),
  tmpdir = tempdir(),
  temp = FALSE,
  outdir = ifelse(temp, tmpdir, dirname(zipped)),
  call = getOption("unzip.call"),
  quiet = TRUE,
  ...
)
}
\arguments{
\item{run}{character (run name)}

\item{path}{path to directory enclosing the NONMEM run directory}

\item{extension}{character: extension to append to run}

\item{filename}{character: supersedes run, extension}

\item{zipped}{character: full path to zipped file (7z extension is optional); supersedes other arguments if supplied}

\item{temp}{whether to unzip to a temporary directory}

\item{outdir}{character: one of \code{path} or \code{tempdir}}

\item{call}{character: system unzip invocation}

\item{tempdir}{character: path to temporary working directory if \code{temp} is true}
}
\value{
invisible path to unzipped file
}
\description{
Unzips a NONMEM file, by default the xml output
of a NONMEM run.  Uses 7z decompression.
}
\examples{
options(nmDir = getOption('qpExampleDir'))
dir <- getOption("qpExampleDir")
unz <- nm_unzip('example1')
unz
unlink(unz)
run <- 'example1'
file <- 'example1.xml'
zipped <- 'example1.xml.7z'
path <- file.path(dir, run, file)
zpath <-file.path(dir, run, zipped)
file.exists(path)
file.exists(zpath)
to <- nm_unzip(zipped = path)
file.exists(to)
unlink(to)
to <- nm_unzip(zipped = zpath)
file.exists(to)
unlink(to)
}
\seealso{
Other nm: 
\code{\link{as_pmx_canonical}()},
\code{\link{definitions.nonmem}()},
\code{\link{extract_varcov}()},
\code{\link{get_est_table.default}()},
\code{\link{get_est_table.nonmem}()},
\code{\link{get_est_table}()},
\code{\link{nm_cov_table}()},
\code{\link{nm_covmat_extract}()},
\code{\link{read_nm.character}()},
\code{\link{read_nm.file}()},
\code{\link{read_nm.run}()},
\code{\link{read_nm.xml}()},
\code{\link{read_nm}()},
\code{\link{sample_uncert}()}
}
\concept{nm}
\keyword{internal}
