% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm_qp.r
\name{read_nm_qp}
\alias{read_nm_qp}
\title{Read NONMEM Results}
\usage{
read_nm_qp(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed}
}
\description{
Reads NONMEM results.
Generic, with method \code{\link{read_nm_qp.character}}
}
\seealso{
Other nm: 
\code{\link{definitions.nonmem}()},
\code{\link{extract_varcov}()},
\code{\link{get_est_table}()},
\code{\link{get_est_table.default}()},
\code{\link{get_est_table.nonmem}()},
\code{\link{get_ofv}()},
\code{\link{get_ofv.character}()},
\code{\link{nm_cov_table}()},
\code{\link{nm_unzip}()},
\code{\link{read_nm_qp.character}()},
\code{\link{read_nm_qp.nonmem}()},
\code{\link{sample_uncert_qp}()}
}
\concept{nm}
\keyword{internal}
