% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_uncert_qp.r
\name{sample_uncert_qp}
\alias{sample_uncert_qp}
\title{Draw Samples from the Quasi-Posterior Parameter Estimate Distribution.}
\usage{
sample_uncert_qp(
  x,
  directory = getOption("nmDir"),
  clean = TRUE,
  quiet = TRUE,
  n = 10,
  seed = 1234,
  est_step = NULL,
  ...
)
}
\arguments{
\item{x}{character string in the form of 'run1'}

\item{directory}{root directory of the run folder where .cov file resides}

\item{clean}{remove the xml and cov file after unzipping? Defaults to TRUE}

\item{quiet}{suppresses the messages (defaults to not suppressing)}

\item{n}{sample size of the requested distribution}

\item{seed}{random seed for simulation}

\item{est_step}{estimation number - defaults to the last one}

\item{...}{ignored}
}
\value{
List of covariance matrices
}
\description{
Read point estimates and the variance-covariance matrix of a model run and create a distribution of parameter estimates forming a sample of the quasi-posterior parameter distribution
}
\examples{

unc1 = sample_uncert_qp("example1", directory = getOption("qpExampleDir"), quiet=TRUE, n=100)
dim(unc1)
}
\seealso{
Other nm: 
\code{\link{definitions.nonmem}()},
\code{\link{extract_varcov}()},
\code{\link{get_est_table}()},
\code{\link{get_est_table.default}()},
\code{\link{get_est_table.nonmem}()},
\code{\link{get_ofv}()},
\code{\link{get_ofv.character}()},
\code{\link{nm_cov_table}()},
\code{\link{nm_unzip}()},
\code{\link{read_nm_qp}()},
\code{\link{read_nm_qp.character}()},
\code{\link{read_nm_qp.nonmem}()}
}
\concept{nm}
